### 2022-10-14 #4856 @editorialbot
[PRE REVIEW]: FijiRelax: Fast and noise-corrected estimation of MRI relaxation maps in 3D + t
**Submitting author:** @Rocsg (<a href="http://orcid.org/0000-0003-3670-044X">Romain Fernandez</a>)
**Repository:** https://github.com/Rocsg/FijiRelax
**Branch with paper.md** (empty if default branch):
**Version:** v4.0.3
