+++ paper.bib
@@ -7,6 +7,7 @@
   pages={129--146},
   year={2012},
   publisher={Wiley Online Library}
+  doi = {10.1111/j.1365-313x.2012.04927.x},
 }
 
 @article{sidi2019multiscale,
@@ -18,6 +19,7 @@
   pages={749--756},
   year={2019},
   publisher={Wiley Online Library}
+  doi = {10.1002/mrc.4905},
 }
 
 @article{han2015drift,
@@ -29,6 +31,7 @@
   pages={1--14},
   year={2015},
   publisher={Springer}
+  doi = {10.1186/s13628-014-0015-1},
 }
 
 @article{van2016quantitative,
@@ -40,6 +43,7 @@
   pages={1176--1188},
   year={2016},
   publisher={American Society of Plant Biologists}
+  doi = {10.1104/pp.15.01388},
 }
 
 @article{GRUSSU2020116884,
@@ -49,9 +53,10 @@ volume = {217},
 pages = {116884},
 year = {2020},
 issn = {1053-8119},
-doi = {https://doi.org/10.1016/j.neuroimage.2020.116884},
+doi = {10.1016/j.neuroimage.2020.116884},
 url = {https://www.sciencedirect.com/science/article/pii/S1053811920303700},
 author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schneider and Julien Cohen-Adad and Timothy M. Shepherd and Daniel C. Alexander and Els Fieremans and Dmitry S. Novikov and Claudia A.M. {Gandini Wheeler-Kingshott}}
+  doi = {10.1016/j.neuroimage.2020.116884},
 }
 
 @dataset{fijirelaxDataset,
@@ -73,6 +78,7 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   pages={1143--1147},
   year={2019},
   organization={IEEE}
+  doi = {10.1109/icip.2019.8804298},
 }
 
 @article{fernandez2021fijiyama,
@@ -84,6 +90,7 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   pages={1482--1484},
   year={2021},
   publisher={Oxford Academic}
+  doi = {10.1093/bioinformatics/btaa846},
 }
 
 @article{gerhards1976analysis,
@@ -95,17 +102,9 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   pages={21--29},
   year={1976},
   publisher={Elsevier}
+  doi = {10.1016/0022-2364(76)90133-5},
 }
 
-@article{grussu2020multi,
-  title={Multi-parametric quantitative in vivo spinal cord MRI with unified signal readout and image denoising},
-  author={Grussu, Francesco and Battiston, Marco and Veraart, Jelle and Schneider, Torben and Cohen-Adad, Julien and Shepherd, Timothy M and Alexander, Daniel C and Fieremans, Els and Novikov, Dmitry S and Wheeler-Kingshott, Claudia AM Gandini},
-  journal={Neuroimage},
-  volume={217},
-  pages={116884},
-  year={2020},
-  publisher={Elsevier}
-}
 
 @article{karakuzu2020qmrlab,
   title={qMRLab: Quantitative MRI analysis, under one umbrella},
@@ -115,6 +114,7 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   number={53},
   pages={2343},
   year={2020}
+  doi = {10.21105/joss.02343},
 }
 
 @article{raya2010t2,
@@ -126,6 +126,7 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   pages={181--193},
   year={2010},
[…]
