+++ paper.bib
@@ -6,8 +6,8 @@
   number={1},
   pages={129--146},
   year={2012},
-  publisher={Wiley Online Library}
-  doi = {10.1111/j.1365-313x.2012.04927.x},
+  publisher={Wiley Online Library},
+  doi = {10.1111/j.1365-313x.2012.04927.x}
 }
 
 @article{sidi2019multiscale,
@@ -18,8 +18,8 @@
   number={9},
   pages={749--756},
   year={2019},
-  publisher={Wiley Online Library}
-  doi = {10.1002/mrc.4905},
+  publisher={Wiley Online Library},
+  doi = {10.1002/mrc.4905}
 }
 
 @article{han2015drift,
@@ -30,8 +30,8 @@
   number={1},
   pages={1--14},
   year={2015},
-  publisher={Springer}
-  doi = {10.1186/s13628-014-0015-1},
+  publisher={Springer},
+  doi = {10.1186/s13628-014-0015-1}
 }
 
 @article{van2016quantitative,
@@ -42,8 +42,8 @@
   number={3},
   pages={1176--1188},
   year={2016},
-  publisher={American Society of Plant Biologists}
-  doi = {10.1104/pp.15.01388},
+  publisher={American Society of Plant Biologists},
+  doi = {10.1104/pp.15.01388}
 }
 
 @article{GRUSSU2020116884,
@@ -55,8 +55,8 @@ year = {2020},
 issn = {1053-8119},
 doi = {10.1016/j.neuroimage.2020.116884},
 url = {https://www.sciencedirect.com/science/article/pii/S1053811920303700},
-author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schneider and Julien Cohen-Adad and Timothy M. Shepherd and Daniel C. Alexander and Els Fieremans and Dmitry S. Novikov and Claudia A.M. {Gandini Wheeler-Kingshott}}
-  doi = {10.1016/j.neuroimage.2020.116884},
+author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schneider and Julien Cohen-Adad and Timothy M. Shepherd and Daniel C. Alexander and Els Fieremans and Dmitry S. Novikov and Claudia A.M. {Gandini Wheeler-Kingshott}},
+  doi = {10.1016/j.neuroimage.2020.116884}
 }
 
 @dataset{fijirelaxDataset,
@@ -77,8 +77,8 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   booktitle={2019 IEEE International Conference on Image Processing (ICIP)},
   pages={1143--1147},
   year={2019},
-  organization={IEEE}
-  doi = {10.1109/icip.2019.8804298},
+  organization={IEEE},
+  doi = {10.1109/icip.2019.8804298}
 }
 
 @article{fernandez2021fijiyama,
@@ -89,8 +89,8 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   number={10},
   pages={1482--1484},
   year={2021},
-  publisher={Oxford Academic}
-  doi = {10.1093/bioinformatics/btaa846},
+  publisher={Oxford Academic},
+  doi = {10.1093/bioinformatics/btaa846}
 }
 
 @article{gerhards1976analysis,
@@ -101,8 +101,8 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   number={1},
   pages={21--29},
   year={1976},
-  publisher={Elsevier}
-  doi = {10.1016/0022-2364(76)90133-5},
+  publisher={Elsevier},
+  doi = {10.1016/0022-2364(76)90133-5}
 }
 
 
[…]
