+++ paper.bib
@@ -125,8 +125,8 @@ author = {Francesco Grussu and Marco Battiston and Jelle Veraart and Torben Schn
   number={1},
   pages={181--193},
   year={2010},
-  publisher={Wiley Online Library}
-  doi = {10.1002/mrm.22178},
+  publisher={Wiley Online Library},
+  doi = {10.1002/mrm.22178}
 }
 
 @article{musse2009monitoring,
