### 2022-10-20 dde878f
Fix method descriptions
