+++ src/main/java/io/github/rocsg/fijirelax/gui/MRI_HyperCurvesExplorer.java
@@ -60,15 +60,8 @@ import io.github.rocsg.fijirelax.mrialgo.HyperMap;
 import io.github.rocsg.fijirelax.mrialgo.MRUtils;
 import io.github.rocsg.fijirelax.mrialgo.RiceEstimator;
 
-// TODO: Auto-generated Javadoc
 /**
- * The Class MRI_HyperCurvesExplorer.
- *
- * @author Rocsg
- * MRI_HyperCurvesExplorer is a user-friendly tool for exploration of T1 T2 relaxation curves coming from T1 and T2 sequence
- * 
- * Acknowledgements :
- * Maida Cardoso and Christophe Goze for data and explanations about T1 and T2 time
+ * MRI_HyperCurvesExplorer is the PlugInFrame describing the curve explorer of FijiRelax, an user-friendly tool for exploration of T1 T2 relaxation curves coming from T1 and T2 sequence
  */
 
 
@@ -419,7 +412,6 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	int curEc2;
 
 /** The gaussian weighting. */
-//	int nRepetMonteCarlo=0;
 	boolean gaussianWeighting=false;
 	
 	/** The max curves. */
@@ -964,8 +956,8 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 
 		
 	/**
-	 *  Entry points and startup functions--------------------------------------------------------------------------------.
-	 *
+	 *  Main function 
+	 *  
 	 * @param args the arguments
 	 */
 	public static void main(String[]args) {
@@ -983,7 +975,7 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	}
 	
 	/**
-	 * Instantiates a new MR I hyper curves explorer.
+	 * Instantiates a new MRI hyper curves explorer.
 	 */
 	public MRI_HyperCurvesExplorer() {
 		super("Vitimage MRI Water tracker ");
@@ -991,7 +983,7 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	}
 	
 	/**
-	 * Run.
+	 * Run the explorer
 	 *
 	 * @param arg the arg
 	 */
@@ -1001,7 +993,7 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	}
 	
 	/**
-	 * Run explorer from hyper map.
+	 * Run explorer by specifying an Hypermap to explore
 	 *
 	 * @param hyp the hyp
 	 */
@@ -1011,9 +1003,9 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	}
 	
 	/**
-	 * Run explorer.
+	 * Run explorer by specifying an HyperMap to explore
 	 *
-	 * @param imgPath the img path
+	 * @param imgPath the path to the HyperMap
 	 */
 	public void runExplorer(String imgPath) {
 		ImagePlus fullHyp=null;
@@ -1023,9 +1015,9 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	}
 		
 	/**
-	 * Run explorer from image.
+	 * Run explorer from a ImagePlus containing an HyperMap
 	 *
-	 * @param fullHyp the full hyp
+	 * @param fullHyp the ImagePlus
 	 */
 	public void runExplorerFromImage(ImagePlus fullHyp) {
 		if(!(fullHyp.getType()==ImagePlus.GRAY32))IJ.run(fullHyp,"32-bit","");
@@ -1064,7 +1056,7 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	
 
 	/**
[…]
