+++ src/main/java/io/github/rocsg/fijirelax/lma/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains linear matrix algebra functions used for the curve fitting algorithms
+ */
+package io.github.rocsg.fijirelax.lma;
+
