+++ src/main/java/io/github/rocsg/fijirelax/curvefit/LMCurveFitterNoBias.java
@@ -1,621 +1,0 @@
-/*
- * 
- */
-package io.github.rocsg.fijirelax.curvefit;
-import io.github.rocsg.fijiyama.common.VitimageUtils;
-import io.github.rocsg.fijirelax.mrialgo.MRUtils;
-import io.github.rocsg.fijirelax.lma.LMAFunction;
-import io.github.rocsg.fijirelax.lma.LMAMatrix.InvertException;
-import io.github.rocsg.fijirelax.lma.LMA;
-
-// TODO: Auto-generated Javadoc
-/**
- * The Class LMCurveFitterNoBias.
- */
-public class LMCurveFitterNoBias {
-	
-	/** The debug LM. */
-	public boolean debugLM=false;
-	
-	/** The Constant timeunits. */
-	public final static String[] timeunits={"ms", "s"};
-	
-	/** The Constant timeitems. */
-	public final static int[] timeitems={MRUtils.MSEC, MRUtils.SEC};
-	
-	/** The Constant fititems2. */
-	public final static String[] fititems2={"Simplex","Levenberg-Marquardt"};
-	
-	/** The Constant constitems2. */
-	public final static int[] constitems2={MRUtils.SIMPLEX,MRUtils.LM};
-	
-	/** The fit. */
-	protected int fit;                // Number of curve type to fit
-	
-	/** The data. */
-	protected double[][] data; // x,y data to fit
-	
-	/** The lma. */
-	private MRLMA lma;
-	
-	/** The parameters. */
-	private double[] parameters;
-	
-	/** The parameters boundaries. */
-	private float[][] parametersBoundaries;
-	
-	/** The gfit. */
-	private float gfit;
-	
-	/** The lambda. */
-	public static double lambda = 0.001;
-	
-	/** The min delta chi 2. */
-	public static double minDeltaChi2 = 1e-30;
-	
-	/** The max iterations. */
-	public static int maxIterations=200;
-
-	
-	
-	 
-	//TODO : after test set, remove from there
-	/**
-	 *  
-	 * Return a first order approximation of initial signal leading to value valFunk, after being corrupted by Rice noise with stddev=sigma .
-	 *
-	 * @param valFunk The observed value after noise alteration
-	 * @param sigma The parameter of the Rice noise
-	 * @return First order approximation of the signal, as a double
-	 */
-	static double sigmaWay(double valFunk,double sigma){
-		double ret=valFunk*valFunk-2*sigma*sigma;
-		return (ret<0 ? 0 : Math.sqrt(ret));
-	}
-	    
-	/**
-	 * Bes funk cost.
-	 *
-	 * @param d the d
-	 * @param sigma2 the sigma 2
-	 * @return the double
-	 */
-	static double besFunkCost(double d,double sigma2) {
-		double alpha=d*d/(4*sigma2*sigma2);
-		return (double)(Math.sqrt(Math.PI*sigma2*sigma2/2.0)*( (1+2*alpha)*bessi0NoExp(alpha) + 2*alpha*bessi1NoExp(alpha) ));
-	}
-
-	/**
-	 * Bes funk deriv.
-	 *
-	 * @param d the d
-	 * @param sigma the sigma
-	 * @return the double
-	 */
-	static double besFunkDeriv(double d,double sigma) {
-		double alpha=d*d/(4*sigma*sigma);
[…]
