+++ src/main/java/io/github/rocsg/fijirelax/curvefit/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains the optimizers and the associated routines for curve fitting T1/T2/PD params using samples of MRI data
+ */
+package io.github.rocsg.fijirelax.curvefit;
+
