+++ src/main/java/io/github/rocsg/fijirelax/mrialgo/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains the data structure description (HyperMap), RiceEstimator functions, and utils for MR data handling
+ */
+package io.github.rocsg.fijirelax.mrialgo;
+

[…]
