### 2022-10-23 5b092ea
Add exhaustive comments (including helpers) Fix a bug under Windows with Curve explorer
