+++ src/main/java/io/github/rocsg/fijirelax/gui/MRI_HyperCurvesExplorer.java
@@ -966,12 +966,11 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	//new MRI_HyperCurvesExplorer().runExplorerFromHyperMap(new HyperMap(IJ.openImage("/home/fernandr/Bureau/Traitements/Sorgho/Series_temporelles/All_timeseries/SSM1_Timeseries.tif")));
 //		new MRI_HyperCurvesExplorer().runExplorerFromHyperMap(new HyperMap(IJ.openImage("/home/fernandr/Bureau/test.tif")));
 //		new MRI_HyperCurvesExplorer().runExplorerFromHyperMap(new HyperMap(IJ.openImage("/home/fernandr/Bureau/test.tif")));
-		ImagePlus img=IJ.openImage("/home/fernandr/Bureau/FijiRelax_DOI/Tutorial_01_First_steps/Input_data/Images/HyperMap_simulated_fantoms.tif");
+		ImagePlus img=IJ.openImage("/home/rfernandez/Bureau/Releases/FijiRelax_DOI/Tutorial_01_First_steps/Input_data/Images/HyperMap_simulated_fantoms.tif");
 //		ImagePlus img=IJ.openImage("/home/fernandr/Bureau/test.tif");
 		HyperMap hyp=new HyperMap(img);
 		hyp.computeMaps();
 		new MRI_HyperCurvesExplorer().runExplorerFromHyperMap(hyp);
-		if(true)return;
 	}
 	
 	/**
@@ -989,6 +988,7 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	 */
 	public void run(String arg) {
 		System.out.println("RUN CALLED !");
+		IJ.log("DebugWIN start 1.1.1");
 		runExplorer(null);
 	}
 	
@@ -1008,9 +1008,12 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	 * @param imgPath the path to the HyperMap
 	 */
 	public void runExplorer(String imgPath) {
+		IJ.log("DebugWIN start 1.1.2 . Given imgPath="+imgPath);
 		ImagePlus fullHyp=null;
 		if (imgPath!=null)fullHyp=IJ.openImage(imgPath);
 		else fullHyp=VitiDialogs.chooseOneImageUI("Open hyperimage", "Open a hyperimage built using T1T2MapImporter");		
+		IJ.log("DebugWIN start 1.1.3 . checking resume of image to show");
+		IJ.log(VitimageUtils.imageResume(fullHyp) );
 		runExplorerFromImage(fullHyp);
 	}
 		
@@ -1020,7 +1023,9 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 	 * @param fullHyp the ImagePlus
 	 */
 	public void runExplorerFromImage(ImagePlus fullHyp) {
+		IJ.log("DebugWIN start 1.1.4 . runExplorerFromImage");
 		if(!(fullHyp.getType()==ImagePlus.GRAY32))IJ.run(fullHyp,"32-bit","");
+		IJ.log("DebugWIN start 1.1.5 . runExplorerFromImage 115");
 		if(VitimageUtils.isBouture(fullHyp))this.deltaT2=25;
 		IJ.log("Opening hyperimage "+VitimageUtils.imageResume(fullHyp));
 		hyperMap=new HyperMap(fullHyp);
@@ -1028,6 +1033,7 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 		this.hasT1=hyperMap.hasT1sequence;
 		this.hasT2=hyperMap.hasT2sequence;
 		
[…]
