### 2022-10-23 3408c52
Clean some debug info
