+++ src/main/java/io/github/rocsg/fijirelax/curvefit/LMDualCurveFitterNoBias.java
@@ -1075,8 +1075,7 @@ public class LMDualCurveFitterNoBias {
 		/**
 		 * Gets estimates of what would be the observed value in absence of Rice noise corruption
 		 *
-		 * @param Tr the recovery time
-		 * @param Te the echo time
+		 * @param x =[Tr,Te]
 		 * @param a the parameters of the equation y=   [ a[0]* Math.exp(-(Te / a[2]))  +  a[1]* Math.exp(-(Te / a[3])) ]
 		 * @return the non riced value
 		 */
@@ -1090,7 +1089,8 @@ public class LMDualCurveFitterNoBias {
 		/**
 		 * Gets estimates of what would be the observed value in absence of Rice noise corruption
 		 *
-		 * @param x =[Tr,Te]
+		 * @param Tr the recovery time
+		 * @param Te the echo time
 		 * @param a the parameters of the equation y=  [ a[0]* Math.exp(-(Te / a[2]))  +  a[1]* Math.exp(-(Te / a[3])) ]
 		 * @return the value in absence of any Rice corruption
 		 */
@@ -1165,8 +1165,7 @@ public class LMDualCurveFitterNoBias {
 		/**
 		 * Gets estimates of what would be the observed value in absence of Rice noise corruption
 		 *
-		 * @param Tr the recovery time
-		 * @param Te the echo time
+		 * @param x =[Tr,Te]
 		 * @param a the parameters of the equation y=   [ a[0]* Math.exp(-(Te / a[2]))  +  a[1]* Math.exp(-(Te / a[3])) ]
 		 * @return the non riced value
 		 */
@@ -1320,7 +1319,7 @@ public class LMDualCurveFitterNoBias {
 	
 	
 	/**
-	 * The Class T1T2Mono. Describes cross-fitting of a T1&T2 relaxation with a single T2 component, without offset and without Rice noise
+	 * The Class T1T2Mono. Describes cross-fitting of a T1 and T2 relaxation with a single T2 component, without offset and without Rice noise
 	 */
 	//////////////////////  4-  T1T2MONO  /////////////////////////////////////////////////////////////////////////////////////////
 	public static class T1T2Mono extends LMAMultiDimFunction {
@@ -1353,7 +1352,8 @@ public class LMDualCurveFitterNoBias {
 		/**
 		 * Gets estimates of what would be the observed value in absence of Rice noise corruption
 		 *
-		 * @param x =[Tr,Te]
+		 * @param Tr the recovery time
+		 * @param Te the echo time
 		 * @param a the parameters of the equation y=  [ a[0]* Math.exp(-(Te / a[2]))* (1 - Math.exp(-(Tr / a[1])) ]
 		 * @return the value in absence of any Rice corruption
 		 */
@@ -1373,8 +1373,7 @@ public class LMDualCurveFitterNoBias {
 		/**
 		 * Gets estimates of what would be the observed value in absence of Rice noise corruption
 		 *
-		 * @param Tr the recovery time
-		 * @param Te the echo time
+		 * @param x =[Tr,Te]
 		 * @param a the parameters of the equation y=   [ a[0]* Math.exp(-(Te / a[2]))  +  a[1]* Math.exp(-(Te / a[3])) ]
 		 * @return the non riced value
 		 */
@@ -1408,7 +1407,7 @@ public class LMDualCurveFitterNoBias {
 	} // end class
 	
 	/**
-	 * The Class T1T2MonoBias. Describes cross-fitting of a T1&T2 relaxation with a single T2 component, with offset and without Rice noise
+	 * The Class T1T2MonoBias. Describes cross-fitting of a T1 and T2 relaxation with a single T2 component, with offset and without Rice noise
 	 */
[…]

[…]
