### 2022-10-24 #4856 @Rocsg
Hello.
I have been looking at the joss_reviewers list to suggest a short-list if that helps. If possible, I would like to suggest 4 reviewers based on their stated programming language and domain/topic preferences:
* xaviercm94
* pr4deepr
* RDastgh1
* MaximLippeveld

I declare that I haven't worked with any of them, nor published with them in the last 5 years.
