### 2022-10-24 #4856 @Rocsg
Another point:
The latest version of FijiRelax is now tagged and released as v4.0.3. It replaces the 4.0.2, which was the "latest" when I submitted to JOSS. This new version (v4.0.3) is now the one pointed by the Fiji updater, and is the latest pointed by Github, Maven Central, Javadoc, and Zenodo.

Best regards
