### 2022-11-03 #4856 @britta-wstnr
Hello @Rocsg! 👋
Thank you for your submission to JOSS!
I see you've already suggested potential reviewers in thread - thanks for that. I will look through your suggestions and send out invitations soon.
