### 2022-11-21 #4856 @Rocsg
Hello britta,
Maybe an ImageJ/Fiji contributor/administrator can provide perspective ? Mark Hiner (hinerm on github) contribute to Fiji, and we never worked together, neither met.
