### 2022-11-24 #4856 @britta-wstnr
Hi @hinerm 👋

Would you be willing to review the submission:
FijiRelax: Fast and noise-corrected estimation of MRI relaxation maps in 3D + t
for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

The review concerns the software [FijiRelax](https://github.com/Rocsg/FijiRelax) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.04856/joss.04856/10.21105.joss.04856.pdf) about it. The review happens fully on GitHub.

You can learn more about the review process and our conflict of interest policy in the reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

If you are available to review this submission, please let me know and I can add you as a reviewer. The review process will start (in a separate GitHub issue) once a sufficient number of reviewers have been found.

Thank you!
Britta
