### 2022-11-25 #4856 @Rocsg
Hi @britta-wstnr
While cross-searching on GitHub / scholar, we identified two other potential reviewers:
* LudgerS is working on MRI software closely related to FijiRelax topics (including denoising MRI relaxation data, and fast exponential fitting).
* matteomancini also works on MRI projects (different types of MRI data).
