### 2022-11-28 #4856 @hinerm
> Would you be willing to review the submission:
> FijiRelax: Fast and noise-corrected estimation of MRI relaxation maps in 3D + t
> for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

Hi @britta-wstnr 👋

Sure, I can be a reviewer.
