### 2022-12-11 #4981 @Rocsg
Hello @britta-wstnr and @kasasxav ,
Here is a point-by-point answer following the integration of requests from Xavier :

#Overview :
Hi,
Thanks for inviting me to review this work [@britta-wstnr](https://github.com/britta-wstnr) and sorry for the delay!
I took a first look into the project, and I can see an extensive documentation including installation details and video tutorials describing the functions of FijiRelax, both in the github repo but also in the ImageJ website. It looks great!
I didn't have the chance to test it yet, but I took a look at the documentation and paper and I have some comments, I believe the author can start with this corrections until I can further test the functionality of the software.

Requests  :

#Question 1 “Insert GPL2 notices in files” :
Documentation
A requirement for GPL-2.0 is to add a file notice at the beginning of each file (see https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html). I see that your files do not have the file notice, so I believe it is needed for a successful review. I also recommend upgrading it to GPL-3.0 although 2.0 is also fine.

Answer :
As requested, we met this requirement by adding a file notice at the beginning of each file, indicating authorship and a reminder of the GPL2.0 licence.

#Question 2 “Integrate Statement of need into the official doc of the project” :
In the github repo you do explain what is the context of the software, however I am missing a clear Statement of Need where you describe throughly what type of problems your tool can solve, for the developer but also the general users. I can see that you have it in the paper, however JOSS requires that you have it in the documentation too - maybe just a smaller and general version of it.

Answer :
We included the Statement of need section to the official ImageJ plugin page, after refactoring this section (see Question 5)

#Question 3  “Document modalities of automatic tests”:
Could you elaborate in how you perform tests? It is a requirement for the review process, however I understand it is hard to do as Github actions as your software is a plugin of ImageJ. Maybe in this case manual testing could be an option? I suggest in this case you could have a separate documentation file in your repo describing different steps to follow to test the software (maybe a checklist?).

Answer :
We added in the repo a file TESTING_AND_RELEASING_HOW-TO.md describing the rationale, how-to and when-to of testing (through maven test) and releasing FijiRelax (jar and code)

#Question 4  “Typos in the README file”:
Some corrections (in github README):
I would add the funding in a separate section "Funding", here you have it in the very beginning (suggestion).
"This tools is developed" -> are
"FijiRelax needs properly formatted dataset" -> article missing, "a properly formatted dataset".
"FijiRelax interface have four main panels" -> has, and also FijiRelax**'s** interface or the interface of FijiRelax.
"Unit for the channels 2 and 3 are milliseconds, what mean you can use it like it" -> meaning or which means.
"(the same, with an additional" -> I would remove "the same"
"This plugin compute" -> computes
When you mention the license, add the version "2.0".

Answer :
Thank you for the careful proof-reading. We included all requested modifications.

#Question 5-a  “Paper structure”:
Paper
You briefly mention how FijiRelax compares to other software in your Statement of Need, however I think it is better to provide a separate section "State of the field" and explain it in more words.
Also when you describe the capabilities, I feel that this shouldn't be in the statement of need either. I would add a section that explains the features of the plugin (or maybe add it in State of the Field pointing out that other software packages do not have those capabilities), and also provide more information about figure 1 to strengthen your claims and help the reader understand the potential of your plugin. I really like that you explain how it can benefit to different users, so I think the best is to go straight to that in your Statement of Need.

Answer :
We proposed a refactoring of the statement of need, including :
A section “Statement of need” going straight to how FijiRelax can benefit to different users.
A section “State of the field”, including a paragraph about benchmarking, introducing capabilities, followed by the list of the main features.
We hope that this modification fits well with the request.

#Question 5-b  “Document scenario for contributors in the repo”:
Paper
I see that you have an API and some directions for developers. I would also suggest adding a sentence about it in the software Documentation, and maybe even in the contributing file.

Answer :
As requested, we copied the paper paragraph indicating the three type of users and their possibilities with Fijiyama, and integrating it as an accessibility disclaim in the CONTRIBUTING.md file

#Question 5-c  “Check the reference order”:
The references are not ordered by appearance in the text. Is there a reason for it? Otherwise I suggest reordering them.

Answer :
We checked that point by checking the refs order in the .bib, by recompiling it, and by checking already published papers in JOSS. From our observations, it seems that this is the result of the automatic compilation using the template of JOSS, making the reference appearing by alphanumeric order.

#Question 6  “Typos in the paper”:
Some corrections
line 17 "dramatically reduces parameters estimation accuracy" -> dramatically reduces the estimation accuracy of the parameters. Or: reduces the parameters**'** estimation accuracy.
In the legend of Figure 1: "results of map computation" and "results of drift artifacts" missing article "the".

Answer : Thank you again for the careful proof-reading. We included all requested modifications.
