### 2022-12-11 707a3d2
Add GPL notice, and edit the doc
