+++ paper.md
@@ -27,21 +27,10 @@ bibliography: paper.bib
 ---
 
 # Summary
-
-Quantitative Magnetic Resonance Imaging (MRI) is a relevant technique to measure water availability and binding in animal or vegetal tissues. This unique capability is useful to decipher the impacts of stresses on living samples [@sidi2019multiscale]. Reproducible measurements of anatomical and physiological changes can be obtained by quantitative mapping of biomarkers (see \autoref{fig:figure1}-a) such as proton density (PD) and magnetization relaxation times (T1, T2) in 2D images [@borisjuk2012surveying ; @van2016quantitative]. However, the application of quantitative MRI to 3D +t microscopic imaging is still challenging. Detection of weak signals associated with small voxels requires longer acquisition times, which favors spatial drift and tissue deformations [@han2015drift]. In addition, low signal-to-noise ratio (SNR) environments dramatically reduce parameters estimation accuracy [@raya2010t2]. Non-specialists facing these issues with multi-echo spin-echo data may be discouraged when using available open implementations of T1-T2 relaxometry: there is a lack of a user-friendly tool capable of processing large data in a reasonable time. We propose FijiRelax, an efficient Fiji plugin for the estimation of relaxation parameter maps from 3D + t multi-echo spin-echo sequences, integrating Rician noise and spatial drift corrections. The plugin is distributed and documented as an ImageJ plugin ([Plugin page](https://imagej.net/plugins/fijirelax)), available in a graphical interface and scriptable in BeanShell to allow batch processing for high-throughput experiments. 
+Quantitative Magnetic Resonance Imaging (MRI) is a relevant technique to measure water availability and binding in animal or vegetal tissues. This unique capability is useful to decipher the impacts of stresses on living samples [@sidi2019multiscale]. Reproducible measurements of anatomical and physiological changes can be obtained by quantitative mapping of biomarkers (see \autoref{fig:figure1}-a) such as proton density (PD) and magnetization relaxation times (T1, T2) in 2D images [@borisjuk2012surveying ; @van2016quantitative]. However, the application of quantitative MRI to 3D +t microscopic imaging is still challenging. Detection of weak signals associated with small voxels requires longer acquisition times, which favors spatial drift and tissue deformations [@han2015drift]. In addition, low signal-to-noise ratio (SNR) environments dramatically reduces the estimation accuracy of the parameters [@raya2010t2]. Non-specialists facing these issues with multi-echo spin-echo data may be discouraged when using available open implementations of T1-T2 relaxometry: there is a lack of a user-friendly tool capable of processing large data in a reasonable time. We propose FijiRelax, an efficient Fiji plugin for the estimation of relaxation parameter maps from 3D + t multi-echo spin-echo sequences, integrating Rician noise and spatial drift corrections. The plugin is distributed and documented as an ImageJ plugin ([Plugin page](https://imagej.net/plugins/fijirelax)), available in a graphical interface and scriptable in BeanShell to allow batch processing for high-throughput experiments. 
 
 # Statement of need
-FijiRelax is a generic tool capable of processing a wide variety of MRI images ranging from a plant stem to a human brain (see \autoref{fig:figure1}-b). It is 100 times faster than open-access implementations of T1-T2 relaxometry curve fitting in MatLab [@karakuzu2020qmrlab] and Python [@GRUSSU2020116884] (see computation time benchmark in \autoref{fig:figure1}-f). FijiRelax provides unique features such as:
-
-* Rice noise integration in the curve fitting algorithm, leading to a more accurate estimation in low SNR situations (see \autoref{fig:figure1}-d)
-
[…]
