+++ src/main/java/io/github/rocsg/fijirelax/gui/Custom_Format_Importer.java
@@ -14,13 +14,24 @@ import ij.plugin.HyperStackConverter;
 import io.github.rocsg.fijirelax.mrialgo.HyperMap;
 import io.github.rocsg.fijirelax.mrialgo.MRUtils;
 
-// TODO: Auto-generated Javadoc
-//TODO : is this class still useful ?
-
 /**
  * The Custom_Format_Importer class is in charge of handling importation MRI data
  * from NIFTI, multiple NIFTI or BRUKER formats.
  *
+ *  	Copyright (C) 2022  <io.github.rocsg>
+ *
+ *   This program is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with this program.  If not, see <https://www.gnu.org/licenses/>
  * @author Romain Fernandez (romain.fernandez@cirad.fr)
  * @version 1.2, 15.04.2022
  */
