+++ src/main/java/io/github/rocsg/fijirelax/curvefit/SimplexDualCurveFitterNoBias.java
@@ -14,7 +14,21 @@ import io.github.rocsg.fijirelax.mrialgo.MRUtils;
 /**
  * A simplex-based solution to curve fitting of exponential functions over MRI observations points
  * This one is preferred as default solution, as it produce way less outliers than Levenberg implementation, is faster, and converges toward the same values in most cases
- */
+ *  	Copyright (C) 2022  <io.github.rocsg>
+ *
+ *   This program is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with this program.  If not, see <https://www.gnu.org/licenses/>
+*/
 public class SimplexDualCurveFitterNoBias{
 	
 	/** The iterations break. */
