+++ src/main/java/io/github/rocsg/fijirelax/gui/MRI_HyperCurvesExplorer.java
@@ -62,7 +62,21 @@ import io.github.rocsg.fijirelax.mrialgo.RiceEstimator;
 
 /**
  * MRI_HyperCurvesExplorer is the PlugInFrame describing the curve explorer of FijiRelax, an user-friendly tool for exploration of T1 T2 relaxation curves coming from T1 and T2 sequence
- */
+ * 	Copyright (C) 2022  <io.github.rocsg>
+ *
+ *   This program is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with this program.  If not, see <https://www.gnu.org/licenses/>
+*/
 
 
 public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListener,KeyListener, MouseListener, MouseWheelListener, MouseMotionListener {/** The init timer. */
