+++ src/main/java/io/github/rocsg/fijirelax/lma/ArrayConverter.java
@@ -1,12 +1,25 @@
 /*
  * 
  */
-package io.github.rocsg.fijirelax.lma;//Initially joalho.data.lma, see  https://zenodo.org/record/4281064
+package io.github.rocsg.fijirelax.lma;
 
-
-// TODO: Auto-generated Javadoc
 /**
  * The Class ArrayConverter.
+ * 	Copyright (C) 2022  <io.github.rocsg>
+ *	//Initially joalho.data.lma, see  https://zenodo.org/record/4281064 , with licence (Open)
+ *
+ *   This program is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with this program.  If not, see <https://www.gnu.org/licenses/>
  */
 public class ArrayConverter {
 	
