### 2022-12-13 Rocsg/FijiRelax#5 @MaximLippeveld
Error when reproducing Figure 1a
**Describe the bug**
When running `Script_figure_1a.bsh` in the directory `FijiRelax_DOI/Experiments_for_figures_and_tables/Scripts` from the test data archive downloaded from Zenodo Ig get the following error:
```Started Script_figure_1a.bsh at Tue Dec 13 09:44:34 CET 2022
javax.script.ScriptException: bsh.EvalError: Sourced file: inline evaluation of: ``import io.github.rocsg.fijirelax.mrialgo.HyperMap; import io.github.rocsg.fijiya . . . '' : Attempt to resolve method: chooseDirectoryUI() on undefined variable or class name: VitiDialogs : at Line: 12 : in file: inline evaluation of: ``import io.github.rocsg.fijirelax.mrialgo.HyperMap; import io.github.rocsg.fijiya . . . '' : VitiDialogs .chooseDirectoryUI ( "Select FijiRelax DOI directory" , "Dir that contains Case_0_Brain, Case_1 ..." )
 in inline evaluation of: ``import io.github.rocsg.fijirelax.mrialgo.HyperMap; import io.github.rocsg.fijiya . . . '' at line number 12
	at bsh.BshScriptEngine.evalSource(BshScriptEngine.java:102)
	at bsh.BshScriptEngine.eval(BshScriptEngine.java:61)
	at javax.script.AbstractScriptEngine.eval(AbstractScriptEngine.java:264)
	at org.scijava.script.ScriptModule.run(ScriptModule.java:164)
	at org.scijava.module.ModuleRunner.run(ModuleRunner.java:163)
	at org.scijava.module.ModuleRunner.call(ModuleRunner.java:124)
	at org.scijava.module.ModuleRunner.call(ModuleRunner.java:63)
	at org.scijava.thread.DefaultThreadService.lambda$wrap$2(DefaultThreadService.java:225)
	at java.util.concurrent.FutureTask.run(FutureTask.java:266)
	at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1149)
	at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:624)
	at java.lang.Thread.run(Thread.java:748)
```

**To Reproduce**
Steps to reproduce the behavior:
1. Open ImageJ
2. Drag script `Script_figure_1a.bsh` on to ImageJ
3. Click on the Run button
4. Get error shown above

**Expected behavior**
Figure 1a is reproduced.

**Screenshots**
If applicable, add screenshots to help explain your problem.

 - OS: Fedora 35
 - ImageJ version: 1.53t
