### 2022-12-13 #4981 @MaximLippeveld
Hi, thanks for the interesting submission! Here are my findings.

I went through the installation process as described on the ImageJ plugin page and everything went smoothly. To test the software I downloaded the dataset from Zenodo and I was able to run Tutorial 2. I ran into a problem when running one of the Figure reproduction scripts and made an issue on the FijiRelax repo: https://github.com/Rocsg/FijiRelax/issues/5

In the paper, I find the second sentence of the summary unclear for a non-specialist audience: "This unique capability is useful to decipher the impacts of stresses on living samples." What unique capability is referenced here? Is it the fact that qMRI can be used to measure water availability and binding or that it can do it in animal or vegetal tissues? The current wording also makes it seem like qMRI is only used to decipher the impacts of stresses on living samples. Is that the case? If not, I would word it like "Among others, MRI is useful to decipher the impacts ..."

Please also provide instructions on how to reproduce the claim of Figure 1f stating that FijiRelax is 100 times quicker compared to other implementations.

Thanks,
Maxim
