### 2022-12-13 Rocsg/FijiRelax#5 @Rocsg
Fixed.
There has been a miss-mess whith the import directives in Script_figure_1a

To fix the problem, one need to insert the following instruction within the import (I inserted it between line 2 and line 3)
import io.github.rocsg.fijiyama.common.VitiDialogs;

The final state, after execution should be this :
![image](https://user-images.githubusercontent.com/24812704/207368818-4c16c07d-dbbe-415f-b4a5-2786721dcd21.png)

I fixed the problem, but currently the Script is versioned with the data in a DOI archive, which is not a convenient way to maintain the test code. Thus I just pushed the beanshell scripts within the FijiRelax Github repository. You can find them in the root/test/Scripts/

I will modify consequently the claims in the README files in both the github repo and the plugin page, and update the README of the data DOI archive in order that users can find these testing Scripts easily. I'll do it before sending my point-by point answer.

Thank you for your feed-back!
