### 2022-12-16 12d1e9d
Clarify the wording of the second sentence
