### 2022-12-17 Rocsg/FijiRelax#6 @kasasxav
GUI not well displayed in Windows 16" screen
**Describe the bug**
A clear and concise description of what the bug is.

**To Reproduce**
Steps to reproduce the behavior:
1. Go to 'Graphical Explorer'

**Expected behavior**
All buttons should be displayed in the GUI.

**Screenshots**

 - OS: Windows
