### 2022-12-17 #4981 @kasasxav
Hi @Rocsg,

thank you for the corrections! I checked some more boxes in the checklist, from the feedback I gave earlier I'm good with your changes.

I also installed the software and tested it with Tutorial 1 and it works smothly, I identified though one bug when running it on my screen and opened an issue (https://github.com/Rocsg/FijiRelax/issues/6). I also plan on going through Tutorial 2, I will let you know if I have further comments on that.

Best,
Xavier.
