+++ paper.md
@@ -39,7 +39,7 @@ FijiRelax is a generic tool for multi-echo spin-echo T1-T2 relaxometry capable o
 * **Developers**: this mode is for programmers fluent with Java and Maven. Start by exploring the FijiRelax API: [API Overview](https://javadoc.io/doc/io.github.rocsg/fijirelax/latest/index.html). Build your own tools on top of the FijiRelax library, provided as a jar file hosted at maven central repository ([Artifact](https://search.maven.org/artifact/io.github.rocsg/fijirelax)), by indicating FijiRelax as a dependency in your POM file and run the unit tests. FijiRelax is hosted on a github public repository ([https://github.com/rocsg/fijirelax](https://github.com/rocsg/fijirelax)) and developers can offer to contribute to its development, and extend it by requesting features, or proposing new features.
 
 # State of the field
-Open-access implementations of T1-T2 relaxometry curve fitting have been released with scripting capabilities in Python [@GRUSSU2020116884] and with a graphical interface in MatLab [@karakuzu2020qmrlab]. FijiRelax offers both scripting capabilities and a graphic interface, and is 100 times faster than these implementations (see computation time benchmark in \autoref{fig:figure1}-f), while including noise-corrected fitting and spatial drift correction. Main features of FijiRelax are:
+Open-access implementations of T1-T2 relaxometry curve fitting have been released with scripting capabilities in Python [@GRUSSU2020116884] and with a graphical interface in MatLab [@karakuzu2020qmrlab]. FijiRelax offers both scripting capabilities and a graphic interface, and is 4 times faster than the Python implementation and 100 times faster than the Matlab implementation (see computation time benchmark in \autoref{fig:figure1}-f), while including noise-corrected fitting and spatial drift correction. Main features of FijiRelax are:
 
 * Rice noise integration in the curve fitting algorithm, leading to a more accurate estimation in low SNR situations (see \autoref{fig:figure1}-d)
 
