+++ test/Python_Scripts/getT2T2star.py
@@ -1,0 +1,677 @@
+### Voxel-wise mono-exponential T2 or T2star fitting on multi-echo spin/gradient echo data
+#
+# Author: Francesco Grussu, University College London
+#		    CDSQuaMRI Project 
+#		   <f.grussu@ucl.ac.uk> <francegrussu@gmail.com>
+#
+# Code released under BSD Two-Clause license
+#
+# Copyright (c) 2019 University College London. 
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
+# 
+# 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
+# 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
+# 
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+# 
+# The views and conclusions contained in the software and documentation are those
+# of the authors and should not be interpreted as representing official policies,
+# either expressed or implied, of the FreeBSD Project.
+
+### Load useful modules
+import argparse, os, sys
+import multiprocessing
+import numpy as np
+from scipy.optimize import minimize
+import nibabel as nib
+import time
+
+def MEsignal(mri_te,tissue_par):
+	''' Generate the signal for a multi-echo experiment at fixed TR
+		
+		
+	    INTERFACE
+	    signal = MEsignal(mri_te,tissue_par)
+	    
+	    PARAMETERS
+	    - mri_te: list/array indicating the TEs (echo times, in ms) used for the experiment (one measurement per TE)
+	    - tissue_par: list/array of tissue parameters, in the following order:
+                          tissue_par[0] = S0 (T1-weighted proton density)
+             		  tissue_par[1] = T2 or T2star (transvere relaxation time, in ms)
+		
+	    RETURNS
+	    - signal: a numpy array of measurements generated according to a multi-echo signal model,
+			
+		         signal  =  S0 * exp(-TE/Txy) 
+		
+		      where TE is the echo time and where S0 and Txy are the tissue parameters (S0 is the T1-weighted proton 
+                      density, and Txy is the transverse relaxation time, i.e. T2 or T2*).
+		
+		
+	    Dependencies (Python packages): numpy
+	    
[…]
