### 2022-12-19 b621c50
Update the paper after Xavier and Maxim comments
