### 2022-12-19 Rocsg/FijiRelax#6 @Rocsg
Thank you @kasasxav for this documented feedback.

At GUI design time, we were maybe too focused on people using large screens, but there is a growing proportion of people working on ultra-portables, in locations that are unpredictable (my advisor do this, as well). I'll try to configure my laptop screen to reproduce this one (it is a matter of #pixels available width X height, I guess), and then publish a patch very soon.
