### 2022-12-19 #4981 @Rocsg
>

Hello @MaximLippeveld . We produce a point-by-point answer to your requests :
Overview :
Hi, thanks for the interesting submission! Here are my findings.
I went through the installation process as described on the ImageJ plugin page and everything went smoothly. To test the software I downloaded the dataset from Zenodo and I was able to run Tutorial 2.

Requests  :

Question 1 “Reproducibility issue with figure 1-a” :
I ran into a problem when running one of the Figure reproduction scripts and made an issue on the FijiRelax repo: [Rocsg/FijiRelax#5](https://github.com/Rocsg/FijiRelax/issues/5)

Answer :
We have fixed the bug and closed the issue. Additionally, we have worked on the versioning of the beanshell scripts. The previous method of storing the scripts in a Zenodo archive was not user-friendly, so we have moved them to the github repository at[ https://github.com/Rocsg/FijiRelax/tree/master/test/Scripts](https://github.com/Rocsg/FijiRelax/tree/master/test/Scripts) . We have also removed the scripts from the Zenodo archive and updated the README files in FijiRelax to explain their location for future users. This will make it easier for users to access the scripts.

Question 2 “Clarification required in the paper” :
In the paper, I find the second sentence of the summary unclear for a non-specialist audience: "This unique capability is useful to decipher the impacts of stresses on living samples." What unique capability is referenced here? Is it the fact that qMRI can be used to measure water availability and binding or that it can do it in animal or vegetal tissues? The current wording also makes it seem like qMRI is only used to decipher the impacts of stresses on living samples. Is that the case? If not, I would word it like "Among others, MRI is useful to decipher the impacts ..."

Answer :
We understand your concern about the wording and have incorporated your suggestion into the paper by stating "Among others, MRI…

Question 3 “How to reproduce performance tests” :
Please also provide instructions on how to reproduce the claim of Figure 1f stating that FijiRelax is 100 times quicker compared to other implementations.

Answer :
As requested, we have made the performance tests reproducible and documented them for publication in JOSS. These tests only measure the computation time for the fitting process and do not include any loading time for the application or data. The original performance tab was created months ago, and since then, MyRelax has implemented multi-threaded computation and eventually other improvements that speeds up the fitting process. As a result, MyRelax is now only 4.6 times slower than FijiRelax, which is a significant improvement. We apologize for our previous statement that MyRelax was 100 times slower and we update the text in the paper to reflect this new information:
“FijiRelax offers both scripting capabilities and a graphic interface, and is 4 times faster than the Python implementation and 100 times faster than the Matlab implementation (see computation time benchmark in \autoref{fig:figure1}-f), while including noise-corrected fitting and spatial drift correction.”

We also decided to include the release number of the benchmarked software to ensure the accuracy of our statements over time.
For QMRLab (the Matlab part), we were unable to find a documented way to multi-thread the process, and the timing is quite unchanged.

Now the instructions for repeating the experiments with FijiRelax, MyRelax and QMRLab can be found in the Github repository, including path to scripts, instructions to reproduce, and release versions of the software used: FijiRelax/REPRODUCING_EXPERIMENTS.md
These modifications are integrated into the new release, patch v4.0.4 - Xavier/Maxim
We appreciate your request for this important investigation, which has had a significant impact on our results and claims.
