+++ src/main/java/io/github/rocsg/fijirelax/gui/MRI_HyperCurvesExplorer.java
@@ -1055,15 +1055,12 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 		echoesText=hyperMap.getEchoesImageText();
 		dims=hyperMap.dims;
 		voxs=hyperMap.voxs;
-		/*if(ij==null)ij = new ImageJ();https://imagej.github.io/
-		else ij = IJ.getInstance();*/
 		IJ.log("Starting MRI Curve Explorer");
 		xCor=hyperMap.dims[0]/2;
 		yCor=hyperMap.dims[1]/2;
 		zCor=0;
 		tCor=0;
 		nTimepoints=hyperMap.T;
-		//setupTimesTrTe();
 		setupStructures();
 		startGui();
 	}
@@ -1460,7 +1457,9 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 		xD=imgCan1.offScreenXD(xMouse);
 		yD=imgCan1.offScreenYD(yMouse);
 		actualizeMriObservationsBasedOnData();
+		System.out.println("Ok 1");
 		computeResultsAgain();			
+		System.out.println("Ok 2");
 		actualizeCursor();
 		displayResultsAgain();
 		updateSwitchButtons();
@@ -1498,8 +1497,8 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 		int screenY=(int) Math.round(screenSize.getHeight());
 		int screenX=(int) Math.round(screenSize.getWidth());
 		if(screenX>1920)screenX=1920;
-		if(screenX<=1600 && false) {
-			IJ.showMessage("Your screen has a very low resolution : "+screenX+" X "+screenY+"\nPlease consider investing in one which have at least 1024 lines x 1480 columns.\nPlugin will run in survivor mode, thus everything can happen");
+		if(screenX<=1600 ) {
+			IJ.showMessage("Your screen has a very low resolution : "+screenX+" X "+screenY+"\nPlugin will run in survivor mode, with reduced polices and graphics");
 			isBionanoDisplay=true;
 		}
 		if(isBionanoDisplay) {screenX=1260; screenY=768;}
@@ -1508,9 +1507,9 @@ public class MRI_HyperCurvesExplorer extends PlugInFrame implements ActionListen
 		TARGET_HEIGHT_FOR_PLOTS_AND_IMGS=isBionanoDisplay ? 233 : 400;
 		TARGET_HEIGHT_FOR_UP_PLOTS_AND_IMGS=isBionanoDisplay ? 280 : 480;
 		TARGET_HEIGHT_FOR_DOWN_PLOT_AND_IMGS=isBionanoDisplay ? 230 : 300;
-		TARGET_WIDTH_FOR_PLOTS_AND_IMGS=isBionanoDisplay ? 235 : 400;
-		TARGET_WIDTH_FOR_PLOTS_T1T2=isBionanoDisplay ? 800 : 1300;
-		TARGET_WIDTH_FOR_PLOTS_ALONE=isBionanoDisplay ? 400 : 650;
+		TARGET_WIDTH_FOR_PLOTS_AND_IMGS=isBionanoDisplay ? 260 : 400;
+		TARGET_WIDTH_FOR_PLOTS_T1T2=isBionanoDisplay ? 900 : 1300;
+		TARGET_WIDTH_FOR_PLOTS_ALONE=isBionanoDisplay ? 450 : 650;
 		DX_IMG=TARGET_WIDTH_FOR_PLOTS_AND_IMGS;
 		DY_IMG=TARGET_HEIGHT_FOR_PLOTS_AND_IMGS;
 		DY_TEXT=isBionanoDisplay ? 18 : 30;
[…]
