### 2022-12-19 Rocsg/FijiRelax#6 @Rocsg
Hello @kasasxav ,
I reproduced your bug using a low resolution mode, then made a fix for the case of small screens. It should activate automatically when you run the explorer with a small screen (width <= 1600). A popup() should appear just before explorer launch to inform that you're running in lowres mode.

To get the release, you can make Fiji > Help > Update, and apply the proposed update (fijirelax.jar)
Let me know if everything is ok!
