### 2022-12-20 #4981 @MaximLippeveld
> Thank you @kasasxav and @MaximLippeveld for making such great progress on the reviewing! star2 A quick reminder that you can also address issues directly in the toolbox' repository (just link to the issues or PRs here as well). As always, feel free to ping me if you have any questions!

@britta-wstnr Should we also address issues relating to the paper as an issue in the the toolbox' repository? Or should those comments be made in this thread?
