### 2022-12-20 Rocsg/FijiRelax#7 @hinerm
Consider adding pom-scijava as a parent pom
Manually specifying dependency versions can lead to skew with what is shipped by Fiji.

**Describe the solution you'd like**
Define a [parent](https://maven.apache.org/guides/introduction/introduction-to-the-pom.html#Project_Inheritance) section in your [pom.xml](https://github.com/Rocsg/FijiRelax/blob/master/pom.xml), pointing to [pom-scijava](https://github.com/scijava/pom-scijava). You should be able to remove most `<version>` tags in your dependencies (for any that are managed by `pom-scijava`), as well as the `<repositories>` block.

The current mechanism of not inheriting from `pom-scijava` is fine, it just requires more maintenance to update component versions, and has a higher chance of not being compatible with some component shipped with Fiji
