### 2022-12-20 Rocsg/FijiRelax#8 @hinerm
Add Support section in README.md
There isn't an immediately obvious path to support, particularly for non-technical users not familiar with GitHub.

**Describe the solution you'd like**
A **Support** section in the `README.md` makes it clear what users should do for technical support. I recommend making a `fiji-relax` tag on [the ImageJ forum](https://forum.image.sc/) and linking to that in the Support section.

With no Support section the implicit default is to use GitHub issues. This is great but may not be clear for non-technical users. The ImageJ forum presents an area for discussion that may not meet the threshold of an "issue", and is also a more centralized option than an individual GitHub repo.
