### 2022-12-20 Rocsg/FijiRelax#9 @hinerm
Unit tests are failing
**Describe the bug**
When I try to build this application on my local system I get several test failures.

**To Reproduce**
Steps to reproduce the behavior:
1. Clone repo
2. Navigate to repo on command line
3. Run `mvn clean test`
4. See errors

**Expected behavior**
The package should build without errors and tests should run after each commit automatically.

**Screenshots**
![image](https://user-images.githubusercontent.com/1123809/208711699-a1486516-3738-40c5-b158-b590871333ca.png)

 - OS: Windows

**Additional context**
For the three `NullPointer` exceptions I believe the problem is that there are [references to local paths](https://github.com/Rocsg/FijiRelax/blob/69a52449b5f5f40db904add714b6fd75cedfd696/src/test/java/io/github/rocsg/fijirelax/test/Test_FijiRelaxPackage.java#L216) in your tests.

Instead, these should be loaded from resources, get via `getClass().getClassLoader().getResources("relative/path/to/resource")`. For example, the following patch fixes `test_08_outliers`:
```
@@ -213,8 +214,7 @@ public class Test_FijiRelaxPackage {
        @Test
        public void test_08_Outliers() throws Exception{
                if(!completeTest)return;
-               String path="/home/rfernandez/eclipse-workspace/MyIJPlugs/FijiRelax/src/test/resources/data/test_2";
-               ImagePlus imgMap=IJ.openImage(new File(path,"TestDataHyperMap/smallSorghoSerieRaw.tif").getAbsolutePath());
+               ImagePlus imgMap=IJ.openImage(new File(getClass().getClassLoader().getResource("data/test_2/TestDataHyperMap/smallSorghoSerieRaw.tif").getFile()).getAbsolutePath());
```

For the `UnsatisfiedLink` error, this is because the platform-specific native library for `SimpleITK` (e.g. `SimpleITKJava.dll` on Windows) is not available. Ideally these would be available through Maven classifiers, but unfortunately I was not able to find them. So instead I would recommend at the start of `test_03_RegisterAndComputeMapsCircle` something simple like trying to create a new `ItkTransform` in a `try...catch` block, and if that fails skipping the test.
