### 2022-12-20 Rocsg/FijiRelax#10 @hinerm
Consider auto-building via GitHub actions
It is nice to know when your commits break your code, so having some sort of continuous integration is always preferable.

**Describe the solution you'd like**
Create a minimal GitHub Action to run the Maven build after each commit to the main brain.

There are other continuous integration options but GH Actions are well-integrated and fairly straightforward to set up.

**Additional context**
This should come after [fixing the unit tests](https://github.com/Rocsg/FijiRelax/issues/9).

Start with the [Building and testing Java with Maven](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-java-with-maven) documentation and adapt the YAML as needed.

Optionally:
- Add a step to grab (e.g. curl) the platform-appropriate `SimpleITKJava` native lib from [the ImageJ-ITK update site](https://sites.imagej.net/ImageJ-ITK/lib/)
- Note that you can use the `setup-java` action to [cache maven dependencies](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-java-with-maven#caching-dependencies)
- Use the [GitHub actions cache](https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows) to cache he `SimpleITKJava` native library
