### 2022-12-20 #4981 @hinerm
Hi @Rocsg!
I'm almost done with my review. I added a couple of issues that I'd like to see resolved:
- Adding a [Support section](https://github.com/Rocsg/FijiRelax/issues/8) in the `README.md`
- Making the [unit tests](https://github.com/Rocsg/FijiRelax/issues/9) work on all systems

I also added two issues that are optional but I thought might make things nicer for you:
- Adding a [parent pom](https://github.com/Rocsg/FijiRelax/issues/7)
- Building with [GitHub acitons](https://github.com/Rocsg/FijiRelax/issues/10)

Other comments:
- On Windows with the JDK that is currently shipped with Fiji (`zulu8.60.0.21-ca-fx-jdk8.0.322-win_x64`) the `ImageJ-ITK` update site is currently broken, because that JDK has some sort of issue loading native libraries (discussion [here](https://forum.image.sc/t/loadlibrary-unsatisfiedlinkerror-with-zulu-openjdk/67997) and [here](https://forum.image.sc/t/imagej-tensorflow-now-with-command-to-change-tensorflow-library-version-switch-to-gpu/31744/23)). This is not your fault at all and has nothing to do with your plugin, but since it could affect your users I wanted you to be aware. The current work-around is to use [a different Java version](https://imagej.net/learn/faq#on-windows).
- I'm not sure I'm seeing the latest [DOI](https://zenodo.org/record/7195985#.Y6HjmHbMKUk) since the date still says Feb 2021... but the `Script_import_and_compute_all_time_series` BeanShell script in `tutorial-02-time-lapse-sorgho` was broken for me. It was just missing two import statements:
```
   import io.github.rocsg.fijiyama.common.VitiDialogs;
   io.github.rocsg.fijiyama.fijiyamaplugin.RegistrationAction;
```
Since that was the script used in [your tutorials](https://imagej.net/plugins/fijirelax) that was the only one I checked, but you might want to double-check others.
