### 2022-12-20 Rocsg/FijiRelax#6 @kasasxav
Still not working sorry :L But my screen is a bit weird, I have a laptop that is 16" (usually they are 15.6").
For me if you could include a way of manually adjusting the size of each widget would work (similar to qt with <>)
