### 2022-12-21 Rocsg/FijiRelax#11 @hinerm
Error reproducing benchmark data
**Describe the bug**
I am trying to run the [table_f.bsh](https://github.com/Rocsg/FijiRelax/blob/master/test/Scripts/Beanshell_Scripts/Script_table_f.bsh) script to reproduce the benchmarking data. I get a `NullPointer` exception in `HyperMap.setupDimensions`.

**To Reproduce**
Steps to reproduce the behavior:
1. Follow instructions to turn on FijiRelax update sites in Fiji
2. Download FijiRelax DOI
3. Open `Script_table_f.bsh` in Fiji
4. Run script
5. Get exception

NB: since this error indicates a null image I assumed the other scripts need to run first to create the input date to `table_f`. When running those scripts I get errors whenever they try to save output, so I'm assuming this is the root of the problem.

**Expected behavior**
Running the script should work. If preconditions (like output from other scripts) are missing, then I should be notified appropriately.

**Screenshots**
Running `Script_table_f.bsh`:
![image](https://user-images.githubusercontent.com/1123809/209006284-c125a99d-f391-401a-bb06-9d74415d1042.png)

Failure from `Script_1b_right.bsh`
![image](https://user-images.githubusercontent.com/1123809/209006496-9dfb00ef-061b-454e-866a-0e18134ffb05.png)

My image output directory after running all scripts:
![image](https://user-images.githubusercontent.com/1123809/209006862-faca1eb4-ad57-4df8-ba9b-837e59e62006.png)

 - OS: Windows 10
