### 2022-12-21 #4981 @hinerm
@Rocsg note that I [ran into some errors](https://github.com/Rocsg/FijiRelax/issues/11) trying to reproduce your benchmark. I will try and take a closer look tomorrow.
