### 2022-12-21 Rocsg/FijiRelax#11 @Rocsg
Hello @hinerm . I am sorry about this one. You are right, there is an image directory missing (I made an archive there with the directory) :
[Human_brain_Hyperimage.zip](https://github.com/Rocsg/FijiRelax/files/10282423/Human_brain_Hyperimage.zip)
.
In the new DOI, this directory (with the image in it) goes into FijiRelax_DOI/Experiments_for_figures_and_tables/Input_data/Images

I had troubles with my internet connection, thus I updated the github repo, but not the DOI (1GB). I'll do it tomorrow morning, including test data refactoring. I keep you informed, meanwhile, you can DL the zip file.

By the way, I am preparing a full answer to your requests. I sincerely thank you, especially for the trick about ImageJ-ITK, I really did not get it right before. I just pushed FijiRelax 4.0.7 including a quick warning for users concerned (using both Windows and zulu), linking to instructions and to your explanation. Many people could be unblocked.
