### 2022-12-22 0d219cc
Fix tests: absolute VS resource path, and JNI excpt.
