### 2022-12-22 83b79dd
Try CI with ubuntu action on push
