### 2022-12-22 b0688c4
Fixed pom setup for mvn clean install (parent+two jama versions)
