### 2022-12-22 bbb06e8
Add mirror insecure exception for vib lib
