+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -7,17 +7,10 @@ jobs:
 
   build:
     runs-on: ubuntu-latest
-
     steps:
       - uses: s4u/maven-settings-action@v2.8.0
-      with:
-        mirrors: |
-          [{
-            "id": "insecure-repo",
-            "mirrorOf": "external:http:*",
-            "url": "http://maven.imagej.net/content/groups/public",
-            "blocked": "false",
-          }]
+        with:
+          mirrors:'[{"id": "insecure-repo", "mirrorOf": "external:http:*", "url": "http://maven.imagej.net/content/groups/public","blocked": "false"}]'
       - uses: actions/checkout@v3
       - name: Set up JDK 11
         uses: actions/setup-java@v3
