### 2022-12-22 3936a98
Add mirror insecure exception for vib lib
