### 2022-12-22 bae8ede
CI with http exception
