+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -5,6 +5,8 @@ on: [push]
 jobs:
   build:
     runs-on: ubuntu-latest
+
+
     steps:
       - uses: s4u/maven-settings-action@v2.8.0
         with:
