+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -5,17 +5,11 @@ on: [push]
 jobs:
   build:
     runs-on: ubuntu-latest
-
-
     steps:
-      - uses: s4u/maven-settings-action@v2.8.0
+      - name Set up settings
+        uses: s4u/maven-settings-action@v2.8.0
         with:
-          mirrors: |
-            [
-              {
-                "id":"insecure-repo"
-              }
-            ]
+          mirrors: '[{"id": "insecure-repo", "name": "httpImageJ", "mirrorOf": "http://maven.imagej.net/content/groups/public", "url": "mirrorUrl"}]'
       - uses: actions/checkout@v3
       - name: Set up JDK 11
         uses: actions/setup-java@v3
