+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -1,16 +1,15 @@
 name: Java CI
-
 on: [push]
-
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-  - name: Set up settings
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Set up settings
         uses: s4u/maven-settings-action@v2.8.0
         with:
           mirrors: '[{"id": "insecure-repo", "name": "httpImageJ", "mirrorOf": "http://maven.imagej.net/content/groups/public", "url": "mirrorUrl"}]'
-      - uses: actions/checkout@v3
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
