+++ .github/workflows/draft-pdf.yml
@@ -1,6 +1,6 @@
 # This is a basic workflow to help you get started with Actions
 
-on: [push]
+on: [tag]
 
 jobs:
   paper:
