+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -9,7 +9,7 @@ jobs:
       - name: Set up settings
         uses: s4u/maven-settings-action@v2.8.0
         with:
-          mirrors: '[{"id": "insecure-repo", "name": "httpImageJ", "mirrorOf": "http://maven.imagej.net/content/groups/public", "url": "http://maven.imagej.net/content/groups/public"}]'
+          mirrors: '[{"id": "insecure-repo", "name": "httpImageJ", "mirrorOf": "external:http:*", "url": "http://maven.imagej.net/content/groups/public", "blocked": "false"}]'
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
