### 2022-12-22 Rocsg/FijiRelax#7 @Rocsg
Hello @hinerm ,
We worked on the exact setup you proposed. It seems to be going well. You will find it in the release 4.0.8
