### 2022-12-22 Rocsg/FijiRelax#8 @Rocsg
Hello @hinerm,
Thank you for this suggestion. We added a support section, following your suggestions. To date, the “support” procedure was to write a mail to me (my personnel mail appear there and there in the GUI). But it seems that it cannot be the single way to get support, and added official support, following your suggestions.
