### 2022-12-22 Rocsg/FijiRelax#9 @Rocsg
Hello @hinerm, we worked on the exact setup you proposed. It seems to be going well. Thank you again. I close the issue
