### 2022-12-22 #4981 @Rocsg
Hello @hinerm ,

Overall, I am grateful for your suggestions, I learned a lot about ImageJ/CI. There is still room to improve my coding patterns and ImageJ lib usage to make me a better contributor to ImageJ ecosystem, I’ll do it on the way.

I implemented your required evolutions. I tried to do the same for your suggestions, but I kept being stuck at some point with github actions. It seems that since maven 3.8, there is a need to “force” maven to authorize gathering dependencies from http: links, which is needed for VIB-lib. I tried to implement a fix in the workflow, but to date, the fix does not work.

In addition, while you did not ask it, It seems to me that using “SNAPSHOT-like” deploying pattern could allow making automatic jar deployment on  push from github to maven (and eventually to sites.imagej.net/Fijiyama). That kind of thing should be really nice to foster community involvement (i.e : automatic sync of maven for each push / accepted pr).

Finally, I used your method to fix the point with ImageJ-ITK which is very nice for our Windows users.

Here is a point-by-point answer to your review, whereas I will answer then to your issues, which is a very convenient way to work.

Overview:
Hi [@Rocsg](https://github.com/Rocsg)!
I'm almost done with my review. I added a couple of issues that I'd like to see resolved:
Question 1:
Adding a [Support section](https://github.com/Rocsg/FijiRelax/issues/8) in the README.md

Answer:
Thank you for this suggestion. We added a support section, following your suggestions. To date, the “support” procedure was to write a mail to me (my personal mail appears here and there in the GUI). But it seems that it cannot be the single way to get support, and added official support, following your suggestions.

Question 2:
Making the [unit tests](https://github.com/Rocsg/FijiRelax/issues/9) work on all systems

Answer:
We worked on the exact setup you proposed. It seems to be going well. Thank you again. I closed the issue

Question 3:
I also added two issues that are optional but I thought might make things nicer for you:
Adding a [parent pom](https://github.com/Rocsg/FijiRelax/issues/7)
Answer:
We worked on the exact setup you proposed. It seems to be going well.

Question 4:
Building with [GitHub acitons](https://github.com/Rocsg/FijiRelax/issues/10)

Answer:
Stuck at a point with a dependence on VIB-lib, requiring to fetch dependence with http: unsecured link (unauthorized since maven 3.8). I keep you in touch.

Question 5:
Other comments:
On Windows with the JDK that is currently shipped with Fiji (zulu8.60.0.21-ca-fx-jdk8.0.322-win_x64) the ImageJ-ITK update site is currently broken, because that JDK has some sort of issue loading native libraries (discussion [here](https://forum.image.sc/t/loadlibrary-unsatisfiedlinkerror-with-zulu-openjdk/67997) and [here](https://forum.image.sc/t/imagej-tensorflow-now-with-command-to-change-tensorflow-library-version-switch-to-gpu/31744/23)). This is not your fault at all and has nothing to do with your plugin, but since it could affect your users I wanted you to be aware. The current workaround is to use [a different Java version](https://imagej.net/learn/faq#on-windows).

Answer:
I really appreciate your help on that point. I reproduced the bug under Windows, and managed a quickfix. I set a showMessage() warning for Windows/Zulu users at the opening of FijiRelax and Fijiyama which refers to the installation guide of FijiRelax, linking to the procedure to follow for using a different Java.

Question 6:
I'm not sure I'm seeing the latest [DOI](https://zenodo.org/record/7195985#.Y6HjmHbMKUk) since the date still says Feb 2021... but the Script_import_and_compute_all_time_series BeanShell script in tutorial-02-time-lapse-sorgho was broken for me. It was just missing two import statements:
  import io.github.rocsg.fijiyama.common.VitiDialogs;
   io.github.rocsg.fijiyama.fijiyamaplugin.RegistrationAction;
Since that was the script used in [your tutorials](https://imagej.net/plugins/fijirelax) that was the only one I checked, but you might want to double-check others.

Answer:
You’re right, the DOI was ancient AND broken. I updated this Wednesday, you can download the new version https://doi.org/10.5281/zenodo.4518730 V5. There is still one script called from the tutorial (that is the one you used. The other ones are not required and not provided anymore. I also changed a bit the setup: as suggested by Xavier, keeping the reproduction scripts in the DOI is not good. I versioned them in Github in test/Scripts/Beanshell_Scripts.
I ran double-check on this. However, if you notice any issues, you can tell me.
