### 2022-12-22 Rocsg/FijiRelax#13 @hinerm
Benchmark reproduction requires silent mode
When reproducing the [benchmark times](https://github.com/Rocsg/FijiRelax/blob/master/REPRODUCING_EXPERIMENTS.md#computation-time-measurement-with-fijirelax), to get a time approximating the claimed 4x speedup I had to [disable the logging calls](https://github.com/Rocsg/FijiRelax/blob/7883c512d2356612d8c11bf51222a862d8c87f54/src/main/java/io/github/rocsg/fijirelax/mrialgo/MRUtils.java#L576-L577).

**Describe the solution you'd like**
An option for "Silent mode" that turns off logging would be nice, given that part of the claim of the project is a performance improvement.

Without a silent mode I got a 3x speedup (2.5s in FijiRelax with logging messages, 2.1s in FijiRelax with logging messages disabled, 7.69s in MyRelax).

**Additional context**
Run on Windows 10 in "high performance" power mode, using Java 1.8.0_252 (Amazon corretto)
