### 2022-12-22 Rocsg/FijiRelax#14 @hinerm
Benchmark times (Windows 10)
I wanted to document my benchmark times. Feel free to close after reviewing.

These tests were run with an Intel i7-9700 CPU @ 3.00GHz, 8 core, 24.0 GB ram on Windows 10.

FijiRelax times out-of-the-box
-------------------------------
* Exp fit: 2.5s
* Offset fit: 3.497s
* Rice fit: 4.076s

FijiRelax times on subsequent runs with [logging disabled](https://github.com/Rocsg/FijiRelax/issues/13)
-----------------------------------------------
* Exp fit: 2.1s
* Offset fit: 3.2s

MyRelax times
---------------
* 7.69s

qMRLab times - MATLAB 2022b
-------------------------
* without_offset: 21.787s
* with_offset: 23.874s

qMRLab times - octvae 7.3.0
-----------------------------
* without_offset: 533.9s
* with_offset: 555.55s
