### 2022-12-22 Rocsg/FijiRelax#16 @hinerm
Consider using SNAPSHOT versioning
Currently your `pom.xml` declares a [release version](https://github.com/Rocsg/FijiRelax/blob/7883c512d2356612d8c11bf51222a862d8c87f54/pom.xml#L43). From a Maven perspective, there can only be one "true" release, but from a Git perspective, many commits now point to the same "release".

**Describe the solution you'd like**
Adding a `-SNAPSHOT` modifier to the version property in your `pom.xml` on non-release commits provides a clear delineation between release and non-release commits, and allows you to leverage the natural Maven machinery for releasing to Release or Snapshot repositories.

You can read more: [here](https://imagej.net/develop/maven#releases-and-snapshots) and [here](https://www.baeldung.com/maven-snapshot-release-repository) as a starting point.

Just incrementing the version when releasing and not using `-SNAPSHOT` is still a valid development style. One can still deduce the commit that corresponds to a release e.g., by looking at the "Implementation-build" in the jar's `MANFIEST.MF`, but it may not be obvious for many, and precludes the ability to deploy intermediate builds between versions.
