### 2022-12-22 #4981 @hinerm
@Rocsg I was able to get the new DOI and [reproduce your times](https://github.com/Rocsg/FijiRelax/issues/14)! Mine were slightly different, but pretty close (my Octave was even slower!). I did have to [turn off IJ.log](https://github.com/Rocsg/FijiRelax/issues/13) to get close to a 4x speedup on my system.

I included Matlab in my tests, which is much faster than Octave but still 10x slower than FijiRelax - but I don't think that needs to be tested in your paper since it's not an open product, right?

Regardless, I opened [a PR](https://github.com/Rocsg/FijiRelax/pull/15) to make your test scripts work in MATLAB, but also not required. Everything worked in Octave out of the box.

>Stuck at a point with a dependence on VIB-lib, requiring to fetch dependence with http: unsecured link (unauthorized since maven 3.8). I keep you in touch

You're so close - you just need to update the [repository block](https://github.com/Rocsg/FijiRelax/blob/7883c512d2356612d8c11bf51222a862d8c87f54/pom.xml#L89) in your pom to use `https`! 😄

>In addition, while you did not ask it, It seems to me that using “SNAPSHOT-like” deploying pattern could allow making automatic jar deployment on push from github to maven (and eventually to sites.imagej.net/Fijiyama). That kind of thing should be really nice to foster community involvement (i.e : automatic sync of maven for each push / accepted pr).

I opened [an issue](https://github.com/Rocsg/FijiRelax/issues/16) to keep track of this - definitely not required for the article though!

@britta-wstnr thanks for making the review process so easy!
