+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -16,5 +16,5 @@ jobs:
           java-version: '11'
           distribution: 'adopt'
       - name: Build with Maven
-        run: mvn --batch-mode --update-snapshots package
+        run: mvn -e --batch-mode --update-snapshots package
 
