+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -16,5 +16,7 @@ jobs:
           java-version: '11'
           distribution: 'adopt'
       - name: Build with Maven
-        run: mvn -e --batch-mode --update-snapshots package
+        run: |
+          export DISPLAY:0.0
+          mvn -e --batch-mode --update-snapshots package
 
