+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -17,6 +17,6 @@ jobs:
           distribution: 'adopt'
       - name: Build with Maven
         run: |
-          export DISPLAY:0.0
+          export DISPLAY=:0.0
           mvn -e --batch-mode --update-snapshots package
 
