+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -17,6 +17,6 @@ jobs:
           distribution: 'adopt'
       - name: Build with Maven
         run: |
-          export DISPLAY=:0.0
-          mvn -e --batch-mode --update-snapshots package
+          export DISPLAY=:99
+          sudo Xvfb -ac :99 screen 0 1280x1024x24 > /dev/null 2>&1 & mvn -e --batch-mode --update-snapshots package
 
