+++ .github/bash/build.sh
@@ -1,0 +1,2 @@
+mvn -e --batch-mode --update-snapshots package
+
