+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -15,8 +15,9 @@ jobs:
         with:
           java-version: '11'
           distribution: 'adopt'
-      - name: Build with Maven
-        run: |
-          export DISPLAY=:99
-          sudo Xvfb -ac :99 screen 0 1280x1024x24 > /dev/null 2>&1 & mvn -e --batch-mode --update-snapshots package
+      - name: Install xvfb
+        run: sudo apt-get install xvfb
+
+      - name: Build with Maven and xvfb
+        run: xvfb-run --autoservernum .github/bash/build.sh
 
