### 2022-12-25 472f42e
Cleaning weird characters for javadoc build
