+++ src/main/java/io/github/rocsg/fijirelax/lma/LMAFunction.java
@@ -20,7 +20,7 @@ import java.util.Arrays;
  *
  *   You should have received a copy of the GNU General Public License
  *   along with this program.  If not, see <https://www.gnu.org/licenses/>
- * @author Janne Holopainen (jaolho@utu.fi, tojotamies@gmail.com)
+ * @author Janne Holopainen (jaolhoATutu.fi, tojotamiesATgmail.com)
  * @version 1.2, 16.11.2006
  */
 public abstract class LMAFunction {
@@ -28,22 +28,22 @@ public abstract class LMAFunction {
 	/**
 	 * Gets the y.
 	 *
-	 * @param x The <i>x</i>-value for which the <i>y</i>-value is calculated.
+	 * @param x The x-value for which the y-value is calculated.
 	 * @param a The fitting parameters.
-	 * @return The <i>y</i>-value of the function.
+	 * @return The y-value of the function.
 	 */
 	public abstract double getY(double x, double[] a);
 	
 	/**
 	 *  
 	 * The method which gives the partial derivates used in the LMA fit.
-	 * If you can't calculate the derivate, use a small <code>a</code>-step (e.g., <i>da</i> = 1e-20)
-	 * and return <i>dy/da</i> at the given <i>x</i> for each fit parameter.
+	 * If you can't calculate the derivate, use a small a-step (e.g., da = 1e-20)
+	 * and return dy/da at the given x for each fit parameter.
 	 *
-	 * @param x The <i>x</i>-value for which the partial derivate is calculated.
+	 * @param x The x-value for which the partial derivate is calculated.
 	 * @param a The fitting parameters.
 	 * @param parameterIndex The parameter index for which the partial derivate is calculated.
-	 * @return The partial derivate of the function with respect to parameter <code>parameterIndex</code> at <i>x</i>.
+	 * @return The partial derivate of the function with respect to parameter parameterIndex at x.
 	 */
 	public abstract double getPartialDerivate(double x, double[] a, int parameterIndex);
 	
