+++ src/main/java/io/github/rocsg/fijirelax/curvefit/LMDualCurveFitterNoBias.java
@@ -21,7 +21,7 @@ import io.github.rocsg.fijirelax.lma.LMA;
  *   GNU General Public License for more details.
  *
  *   You should have received a copy of the GNU General Public License
- *   along with this program.  If not, see <https://www.gnu.org/licenses/>
+ *   along with this program.  If not, see https://www.gnu.org/licenses/
 */
 public class LMDualCurveFitterNoBias {	
 	
