+++ src/main/java/io/github/rocsg/fijirelax/curvefit/package-info.java
@@ -1,7 +1,7 @@
 /**
  * This package contains the optimizers and the associated routines for curve fitting T1/T2/PD params using samples of MRI data
  *  	Copyright (C) 2022  <io.github.rocsg> (except MRLMA class, which is 
- *  	@author Janne Holopainen (jaolho@utu.fi, tojotamies@gmail.com)
+ *  	@author Janne Holopainen (jaolhoATutu.fi, tojotamiesATgmail.com)
  *   @version 1.2, 24.04.2007 , free for non-commercial use)
  *
  *
@@ -16,7 +16,7 @@
  *   GNU General Public License for more details.
  *
  *   You should have received a copy of the GNU General Public License
- *   along with this program.  If not, see <https://www.gnu.org/licenses/>
+ *   along with this program.  If not, see https://www.gnu.org/licenses/
  */
 package io.github.rocsg.fijirelax.curvefit;
 
