### 2022-12-25 1e6b35f
Cleaning weird characters for javadoc build
