+++ src/main/java/io/github/rocsg/fijirelax/gui/Custom_Format_Importer.java
@@ -34,7 +34,7 @@ import io.github.rocsg.fijirelax.mrialgo.MRUtils;
  *   GNU General Public License for more details.
  *
  *   You should have received a copy of the GNU General Public License
- *   along with this program.  If not, see <https://www.gnu.org/licenses/>
+ *   along with this program.  If not, see https://www.gnu.org/licenses/
  * @author Romain Fernandez (romain.fernandezATcirad.fr)
  * @version 1.2, 15.04.2022
  */
