+++ src/main/java/io/github/rocsg/fijirelax/curvefit/package-info.java
@@ -1,6 +1,6 @@
 /**
  * This package contains the optimizers and the associated routines for curve fitting T1/T2/PD params using samples of MRI data
- *  	Copyright (C) 2022  <io.github.rocsg> (except MRLMA class, which is 
+ *  	Copyright (C) 2022  io.github.rocsg (except MRLMA class, which is 
  *  	@author Janne Holopainen (jaolhoATutu.fi, tojotamiesATgmail.com)
  *   @version 1.2, 24.04.2007 , free for non-commercial use)
  *
