+++ src/main/java/io/github/rocsg/fijirelax/lma/LMAMultiDimFunction.java
@@ -21,9 +21,9 @@ import java.util.Arrays;
  *   GNU General Public License for more details.
 
  *   You should have received a copy of the GNU General Public License
- *   along with this program.  If not, see <https://www.gnu.org/licenses/>
+ *   along with this program.  If not, see https://www.gnu.org/licenses/
  * 
- * @author Janne Holopainen (jaolho@utu.fi, tojotamies@gmail.com)
+ * @author Janne Holopainen (jaolhoATutu.fi, tojotamiesATgmail.com)
  * @version 1.0, 23.03.2007
  */
 public abstract class LMAMultiDimFunction {
