+++ src/main/java/io/github/rocsg/fijirelax/lma/LMAFunction.java
@@ -19,7 +19,7 @@ import java.util.Arrays;
  *   GNU General Public License for more details.
  *
  *   You should have received a copy of the GNU General Public License
- *   along with this program.  If not, see <https://www.gnu.org/licenses/>
+ *   along with this program.  If not, see https://www.gnu.org/licenses/
  * @author Janne Holopainen (jaolhoATutu.fi, tojotamiesATgmail.com)
  * @version 1.2, 16.11.2006
  */
