+++ src/main/java/io/github/rocsg/fijirelax/curvefit/MRLMA.java
@@ -45,7 +45,7 @@ public class MRLMA extends LMA{
 	public boolean verbose = false;
 	/** 
 	 * The model function to be fitted, y = y(x[], a[]),
-	 * where <code>x[]</code> the array of x-values and <code>a</code>
+	 * where x[] the array of x-values and a
 	 * is the array of fit parameters.
 	 */
 	public LMAMultiDimFunction function;
@@ -135,15 +135,15 @@ public class MRLMA extends LMA{
 	/**
 	 * One dimensional convenience constructor for LMAFunction.
 	 * You can also implement the same function using LMAMultiDimFunction.
-	 * <p>
+	 * 
 	 * Initiates the fit with function constructed weights and a JAMA matrix.
 	 * N is the number of data points, M is the number of fit parameters.
-	 * Call <code>fit()</code> to start the actual fitting.
+	 * Call fit() to start the actual fitting.
 	 *
 	 * @param function The model function to be fitted. Must be able to take M input parameters.
 	 * @param parameters The initial guess for the fit parameters, length M.
-	 * @param dataPoints The data points in an array, <code>double[0 = x, 1 = y][point index]</code>.
-	 * Size must be <code>double[2][N]</code>.
+	 * @param dataPoints The data points in an array, double[0 = x, 1 = y][point index].
+	 * Size must be double[2][N].
 	 * @param bounds the bounds
 	 */
 	public MRLMA(final LMAFunction function, double[] parameters, double[][] dataPoints, float[][]bounds) {
@@ -154,15 +154,15 @@ public class MRLMA extends LMA{
 	/**
 	 * One dimensional convenience constructor for LMAFunction.
 	 * You can also implement the same function using LMAMultiDimFunction.
-	 * <p>
+	 * 
 	 * Initiates the fit with function constructed weights and a JAMA matrix.
 	 * N is the number of data points, M is the number of fit parameters.
-	 * Call <code>fit()</code> to start the actual fitting.
+	 * Call fit() to start the actual fitting.
 	 *
 	 * @param function The model function to be fitted. Must be able to take M input parameters.
 	 * @param parameters The initial guess for the fit parameters, length M.
-	 * @param dataPoints The data points in an array, <code>double[0 = x, 1 = y][point index]</code>.
-	 * Size must be <code>double[2][N]</code>.
+	 * @param dataPoints The data points in an array, double[0 = x, 1 = y][point index].
+	 * Size must be double[2][N].
 	 * @param weights the weights
 	 * @param bounds the bounds
 	 */
@@ -193,15 +193,15 @@ public class MRLMA extends LMA{
 	/**
 	 * One dimensional convenience constructor for LMAFunction.
 	 * You can also implement the same function using LMAMultiDimFunction.
-	 * <p>
+	 * 
 	 * Initiates the fit with function constructed weights and a JAMA matrix.
 	 * N is the number of data points, M is the number of fit parameters.
-	 * Call <code>fit()</code> to start the actual fitting.
+	 * Call fit() to start the actual fitting.
 	 *
 	 * @param function The model function to be fitted. Must be able to take M input parameters.
 	 * @param parameters The initial guess for the fit parameters, length M.
-	 * @param dataPoints The data points in an array, <code>float[0 = x, 1 = y][point index]</code>.
-	 * Size must be <code>float[2][N]</code>.
+	 * @param dataPoints The data points in an array, float[0 = x, 1 = y][point index].
+	 * Size must be float[2][N].
 	 * @param bounds the bounds
 	 */
[…]
