+++ .github/workflows/draft-pdf.yml
@@ -1,6 +1,9 @@
 # This is a basic workflow to help you get started with Actions
 
-on: [push]
+on:
+  push:
+    paths:
+      - paper.md
 
 jobs:
   paper:
