+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -16,7 +16,7 @@ jobs:
           apt-get install curl -y
       - name: grab SimpleITK
         run: curl -L https://sites.imagej.net/ImageJ-ITK/lib/linux64/libSimpleITKJava.so-20150910103835
-              - name: Set up JDK 11
+      - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
           java-version: '11'
