+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -10,12 +10,10 @@ jobs:
         uses: s4u/maven-settings-action@v2.8.0
         with:
           mirrors: '[{"id": "imagej.public", "name": "imagej.public", "mirrorOf": "external:http:*", "url": "http://maven.imagej.net/content/groups/public", "blocked": "false"}]'
-      - name: install curl
-        run: |
-          apt-get update
-          apt-get install curl -y
       - name: grab SimpleITK
-        run: curl -L https://sites.imagej.net/ImageJ-ITK/lib/linux64/libSimpleITKJava.so-20150910103835
+        uses: wei/curl@master
+          with:
+            args: https://sites.imagej.net/ImageJ-ITK/lib/linux64/libSimpleITKJava.so-20150910103835 -o libSimpleITKJava.so
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
