+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -12,8 +12,8 @@ jobs:
           mirrors: '[{"id": "imagej.public", "name": "imagej.public", "mirrorOf": "external:http:*", "url": "http://maven.imagej.net/content/groups/public", "blocked": "false"}]'
       - name: grab SimpleITK
         uses: wei/curl@master
-          with:
-            args: https://sites.imagej.net/ImageJ-ITK/lib/linux64/libSimpleITKJava.so-20150910103835 -o libSimpleITKJava.so
+        with:
+          args: https://sites.imagej.net/ImageJ-ITK/lib/linux64/libSimpleITKJava.so-20150910103835 -o libSimpleITKJava.so
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
