+++ lib/README.md
@@ -1,0 +1,2 @@
+This directory is there to store temporary native libraries at test time
+
