+++ .github/bash/build.sh
@@ -1,2 +1,2 @@
-mvn -e --batch-mode --update-snapshots package -Djava.library.path=lib
+mvn -e --batch-mode --update-snapshots package -Djava.library.path="/home/runner/work/FijiRelax/FijiRelax/lib"
 
