+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -13,7 +13,7 @@ jobs:
       - name: grab SimpleITK
         uses: wei/curl@master
         with:
-          args: https://sites.imagej.net/ImageJ-ITK/lib/linux64/libSimpleITKJava.so-20150910103835 -o lib/libSimpleITKJava.so
+          args: https://sites.imagej.net/ImageJ-ITK/lib/linux64/libSimpleITKJava.so-20150910103835 -o /lib/libSimpleITKJava.so
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
