### 2023-01-02 Rocsg/FijiRelax#10 @Rocsg
Hello @hinerm ,
I wish you a fresh and cool new year, with what you want in it.

I think I struggled to setup curl grabbing, and caching dependancies, until the test with registration finally starts and execute well. That is really nice, I bet I could setup such tests for Fijiyama, too.

For the native library, I struggled until it worked, and finished with it in my pom:

	          <artifactId>maven-surefire-plugin</artifactId>
              <configuration>
                  <argLine>-Djava.library.path=${java.library.path};/home/runner/work/FijiRelax/FijiRelax/lib/</argLine>
              </configuration>
	        </plugin>

I did that because I was stuck with the impossibility to change dynamically java.library.path in the mvn command activated by the action. I guess there should be more elegant ways to do this. However, that pom can be used both in github action and on a local config (reacts well with mvn install).
