### 2023-01-02 ac764ff
Add a silent mode for perfs
