+++ src/main/java/io/github/rocsg/fijirelax/mrialgo/HyperMap.java
@@ -1644,7 +1644,7 @@ public class HyperMap {
 	 * @return the updated HyperMap, as a TIFF ImagePlus
 	 */	
 	public ImagePlus computeMaps() {
-		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,NoiseManagement.RICE,false,null,4);
+		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,NoiseManagement.RICE,false,null,4,false);
 	}
 
 	/**
@@ -1654,7 +1654,7 @@ public class HyperMap {
 	 * @return the updated HyperMap, as a TIFF ImagePlus
 	 */
 	public ImagePlus computeMaps(NoiseManagement noise) {
-		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,noise,false,null,4);
+		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,noise,false,null,4,false);
 	}
 
 	/**
@@ -1663,7 +1663,7 @@ public class HyperMap {
 	 * @return the updated HyperMap, as a TIFF ImagePlus
 	 */	
 	public ImagePlus computeMapsNoJoint() {
-		return computeMapsAgainAndMask(MRUtils.SIMPLEX,true,NoiseManagement.RICE,false,null,4);
+		return computeMapsAgainAndMask(MRUtils.SIMPLEX,true,NoiseManagement.RICE,false,null,4,false);
 	}
 	
 	/**
@@ -1673,7 +1673,7 @@ public class HyperMap {
 	 * @return the updated HyperMap, as a TIFF ImagePlus
 	 */
 	public ImagePlus computeMapsNoJoint(NoiseManagement noi) {
-		return computeMapsAgainAndMask(MRUtils.SIMPLEX,true,noi,false,null,4);
+		return computeMapsAgainAndMask(MRUtils.SIMPLEX,true,noi,false,null,4,false);
 	}
 
 	/**
@@ -1683,7 +1683,7 @@ public class HyperMap {
 	 * @return the updated HyperMap, as a TIFF ImagePlus
 	 */
 	public ImagePlus computeMapsWithMask(ImagePlus mask) {
-		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,NoiseManagement.RICE,false,mask,4);
+		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,NoiseManagement.RICE,false,mask,4,false);
 	}
 	
 	/**
@@ -1693,7 +1693,7 @@ public class HyperMap {
 	 * @return the updated HyperMap, as a TIFF ImagePlus
 	 */
 	public ImagePlus computeMaps(double nStdDevForMask) {
-		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,NoiseManagement.RICE,false,null,nStdDevForMask);
+		return computeMapsAgainAndMask(MRUtils.SIMPLEX,false,NoiseManagement.RICE,false,null,nStdDevForMask,false);
 	}
 
 	/**
@@ -1711,7 +1711,7 @@ public class HyperMap {
 		if(paramsGUI[1]==0)noi=NoiseManagement.RICE;
 		else noi=(paramsGUI[1]==1 ? NoiseManagement.OFFSET : NoiseManagement.NOTHING);
 		double nbStd=paramsGUI[4];
-		return computeMapsAgainAndMask(algType,separated ,noi,forgetFistEcho,imgMask,nbStd);
+		return computeMapsAgainAndMask(algType,separated ,noi,forgetFistEcho,imgMask,nbStd,false);
 	}
 
 	/**
@@ -1725,16 +1725,17 @@ public class HyperMap {
[…]
