### 2023-01-02 Rocsg/FijiRelax#13 @Rocsg
Hello @hinerm ,
I add this feature, and updated the test scripts accordingly.

Another point I see that you are testing under Windows, and I estimate that there could be a way to setup automatic testing with github action and Windows. I set an issue about it and will plan to work on it.

Romain
