### 2023-01-02 Rocsg/FijiRelax#16 @Rocsg
Hi @hinerm ,
I setup this new behaviour for both Fijiyama (4.0.7-SNAPSHOT) and FijiRelax (4.0.9-SNAPSHOT). I bet I can go on on snapshots for a while, and when wanting to "fix" things, can deploy and tag releases 4.0.7 and 4.0.9 , then go on working, but on 4.0.8-SNAPSHOT and 4.0.10-SNAPSHOT

Another thing, now that maven repository can be "overwritten" at any time, I should consider to automatically push to maven rep and sites.imagej.net/Fijiyama  from github action. I need to learn more about it, to estimate the "balance" in term of client risk. I set an issue for it.

Thank you again.
