+++ .github/workflows/CI_windows_on_push.yml
@@ -1,0 +1,25 @@
+name: Java CI
+on: [push]
+jobs:
+  build:
+    runs-on: windows-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Set up settings
+        uses: s4u/maven-settings-action@v2.8.0
+        with:
+          mirrors: '[{"id": "imagej.public", "name": "imagej.public", "mirrorOf": "external:http:*", "url": "http://maven.imagej.net/content/groups/public", "blocked": "false"}]'
+      - name: grab SimpleITK
+        uses: wei/curl@master
+        with:
+          args: https://sites.imagej.net/ImageJ-ITK/lib/win64/SimpleITKJava.dll-20150910103835 -o lib/libSimpleITKJava.dll
+      - name: Set up JDK 11
+        uses: actions/setup-java@v3
+        with:
+          java-version: '11'
+          distribution: 'adopt'
+          cache: maven
+      - name: Build with Maven
+        run: mvn -e --batch-mode --update-snapshots package -Djava.library.path="/home/runner/work/FijiRelax/FijiRelax/lib"
+
