+++ .github/workflows/CI_windows_on_push.yml
@@ -1,4 +1,4 @@
-name: Java CI
+name: Java CI WINDOWS
 on: [push]
 jobs:
   build:
@@ -11,9 +11,9 @@ jobs:
         with:
           mirrors: '[{"id": "imagej.public", "name": "imagej.public", "mirrorOf": "external:http:*", "url": "http://maven.imagej.net/content/groups/public", "blocked": "false"}]'
       - name: grab SimpleITK
-        uses: wei/curl@master
+        uses: wei/wget@v1
         with:
-          args: https://sites.imagej.net/ImageJ-ITK/lib/win64/SimpleITKJava.dll-20150910103835 -o lib/libSimpleITKJava.dll
+          args: -O lib/libSimpleITKJava.dll -P https://sites.imagej.net/ImageJ-ITK/lib/win64/SimpleITKJava.dll-20150910103835 
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
