+++ .github/workflows/CI_windows_on_push.yml
@@ -13,7 +13,7 @@ jobs:
       - name: grab SimpleITK
         uses: wei/wget@v1
         with:
-          args: -O lib/libSimpleITKJava.dll -P https://sites.imagej.net/ImageJ-ITK/lib/win64/SimpleITKJava.dll-20150910103835 
+          args: https://sites.imagej.net/ImageJ-ITK/lib/win64/SimpleITKJava.dll-20150910103835 -O lib/libSimpleITKJava.dll
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
