+++ .github/workflows/CI_windows_on_push.yml
@@ -20,5 +20,5 @@ jobs:
           cache: maven
       - name: Build with Maven
         run: mvn -e --batch-mode --update-snapshots package 
-          options: -Djava.library.path="D:\a\FijiRelax\FijiRelax\lib" -Djava.awt.headless=true
+          options: -Djava.awt.headless=true
 
