+++ .github/bash/buildWin.sh
@@ -1,0 +1,2 @@
+mvn -e --batch-mode --update-snapshots package -Djava.library.path="D:\a\FijiRelax\FijiRelax\lib" -Djava.awt.headless=true
+
