+++ .github/workflows/CI_windows_on_push.yml
@@ -19,5 +19,7 @@ jobs:
           distribution: 'adopt'
           cache: maven
       - name: Build with Maven
-        run: .github/bash/buildWin.sh
+        run: |
+          mvn -e --batch-mode --update-snapshots package
+        options: -Djava.library.path="D:\a\FijiRelax\FijiRelax\lib" -Djava.awt.headless=true
 
