+++ .github/workflows/CI_windows_on_push.yml
@@ -21,5 +21,5 @@ jobs:
       - name: Build with Maven
         run: |
           mvn -e --batch-mode --update-snapshots package
-        options: -Djava.library.path="D:\a\FijiRelax\FijiRelax\lib" -Djava.awt.headless=true
+        args: -Djava.library.path="D:\a\FijiRelax\FijiRelax\lib" -Djava.awt.headless=true
 
