+++ .github/workflows/CI_windows_on_push.yml
@@ -19,7 +19,6 @@ jobs:
           distribution: 'adopt'
           cache: maven
       - name: Build with Maven
-        run: |
-          mvn -e --batch-mode --update-snapshots package
+        run: mvn -e --batch-mode --update-snapshots package
         args: -Djava.library.path="D:\a\FijiRelax\FijiRelax\lib" -Djava.awt.headless=true
 
