+++ .github/bash/buildWin.sh
@@ -1,3 +1,3 @@
 echo "Toto"
-mvn -e --batch-mode --update-snapshots package -Djava.library.path="D:\a\FijiRelax\FijiRelax\lib" -Djava.awt.headless=true
+
 
