+++ .github/workflows/CI_windows_on_push.yml
@@ -11,7 +11,7 @@ jobs:
         with:
           mirrors: '[{"id": "imagej.public", "name": "imagej.public", "mirrorOf": "external:http:*", "url": "http://maven.imagej.net/content/groups/public", "blocked": "false"}]'
       - name: grab SimpleITK
-        run: Invoke-WebRequest https://sites.imagej.net/ImageJ-ITK/lib/win64/SimpleITKJava.dll-20150910103835 -OutFile lib/libSimpleITKJava.dll
+        run: Invoke-WebRequest https://sites.imagej.net/ImageJ-ITK/lib/win64/SimpleITKJava.dll-20150910103835 -OutFile lib/SimpleITKJava.dll
       - name: Set up JDK 11
         uses: actions/setup-java@v3
         with:
