+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -1,5 +1,5 @@
 name: Java CI LINUX
-on: [push]
+on: [pull_request]
 jobs:
   build:
     runs-on: ubuntu-latest
