+++ .github/workflows/CI_windows_on_push.yml
@@ -21,5 +21,5 @@ jobs:
       - name: Build with Maven
         shell: bash
         run: |
-          exec mvn -e --batch-mode --update-snapshots package -Djava.awt.headless=true -Djava.library.path=D:\a\FijiRelax\FijiRelax\lib\
+          exec mvn -e --batch-mode --update-snapshots package -Djava.awt.headless=true -Djava.library.path=D:/a/FijiRelax/FijiRelax/lib/
 
