+++ src/main/java/io/github/rocsg/fijirelax/mrialgo/HyperMap.java
@@ -1984,6 +1984,7 @@ public class HyperMap {
 			this.setDisplayRange();
 		}
 		adjustContrast();
+		
 	}
 		
 	/**
