+++ .github/workflows/CI_windows_on_push.yml
@@ -17,7 +17,6 @@ jobs:
         with:
           java-version: '11'
           distribution: 'adopt'
-          cache: maven
       - name: Build with Maven
         shell: bash
         run: |
