### 2023-01-03 5b18892
Ease auto tests headless
