### 2023-01-03 e24225c
Ease auto tests headless
