+++ .github/workflows/CI_ubuntu_on_push.yml
@@ -1,5 +1,5 @@
 name: Java CI LINUX
-on: [pull_request]
+on: [push]
 jobs:
   build:
     runs-on: ubuntu-latest
