### 2023-01-16 #4981 @britta-wstnr
Thanks for confirming, @kasasxav, @MaximLippeveld, and @hinerm!

@Rocsg I went over the pdf and opened a PR to fix some typos. Please make sure to check the pdf yourself as well for things you might want to fix before the publication. The formal things (like references) seem to be okay.

Once the pdf is in shape, we can move on!
