### 2023-01-17 7450b71
Proof typos correction by author #1
