+++ paper.md
@@ -15,9 +15,9 @@ authors:
     orcid: 0000-0003-4637-4713
     affiliation: 3
 affiliations:
- - name: CIRAD, UMR AGAP, F-34398 Montpellier, France
+ - name: CIRAD, UMR AGAP Institut, F-34398 Montpellier, France
    index: 1
- - name: AGAP, Univ Montpellier, CIRAD, INRAE, Institut Agro, Montpellier, France.
+ - name: AGAP Institut, Univ Montpellier, CIRAD, INRAE, Institut Agro, Montpellier, France.
    index: 2
  - name: IFV, French Institute of Vine and Wine, UMT Géno-Vigne, UMR AGAP, F-34398 Montpellier, France.
    index: 3
@@ -30,16 +30,16 @@ bibliography: paper.bib
 Quantitative Magnetic Resonance Imaging (MRI) is a relevant technique to measure water availability and binding in animal or vegetal tissues. Among others, MRI is useful to decipher the impacts of stresses on living samples [@sidi2019multiscale]. Reproducible measurements of anatomical and physiological changes can be obtained by quantitative mapping of biomarkers (see \autoref{fig:figure1}-a) such as proton density (PD) and magnetization relaxation times (T1, T2) in 2D images [@borisjuk2012surveying ; @van2016quantitative]. However, the application of quantitative MRI to 3D + t microscopic imaging is still challenging. Detection of weak signals associated with small voxels requires longer acquisition times, which favors spatial drift and tissue deformations [@han2015drift]. In addition, low signal-to-noise ratio (SNR) environments dramatically reduce the estimation accuracy of the parameters [@raya2010t2]. Non-specialists facing these issues with multi-echo spin-echo data may be discouraged when using available open implementations of T1-T2 relaxometry: there is a lack of a user-friendly tools capable of processing large data in a reasonable time. We propose FijiRelax, an efficient Fiji plugin for the estimation of relaxation parameter maps from 3D + t multi-echo spin-echo sequences, integrating Rician noise and spatial drift corrections. The plugin is distributed and documented as an ImageJ plugin ([Plugin page](https://imagej.net/plugins/fijirelax)), available in a graphical interface and scriptable in BeanShell to allow batch processing for high-throughput experiments. 
 
 # Statement of need
-FijiRelax is a generic tool for multi-echo spin-echo T1-T2 relaxometry capable of processing a wide variety of MRI images ranging from a plant stem to a human brain (see \autoref{fig:figure1}-b). It has been designed for three types of scientists: i) end-users using a GUI, ii) advanced users able to use a scripting languages to process large number of images, and iii) developers able to adapt and extend the application with new functionalities.
+FijiRelax is a generic tool for multi-echo spin-echo T1-T2 relaxometry capable of processing a wide variety of MRI images ranging from a plant stem to a human brain (see \autoref{fig:figure1}-b). It has been designed for three types of scientists: i) end-users using a GUI, ii) advanced users able to use a scripting language to process large number of images, and iii) developers able to adapt and extend the application with new functionalities.
 
[…]
