### 2023-01-20 #4981 @Rocsg
Hello @britta-wstnr ,
- I made the v4.0.10 tagged release :
https://github.com/Rocsg/FijiRelax/releases/tag/v4.0.10

- I created a Zenodo archive of that release :
https://zenodo.org/record/7554691#.Y8qjRdLMJhE

- I fixed the info of the Zenodo repo:
Title: "FijiRelax: Fast and noise-corrected estimation of MRI relaxation maps in 3D + t"
Authors : Romain Fernandez (+ORCID) and Cedric Moisy (+ORCID).

- I posted here the release version number: 4.0.10, and the links (just above).

Is everything okay?

Romain
