+++ paper.md
@@ -34,9 +34,9 @@ FijiRelax is a generic tool for multi-echo spin-echo T1-T2 relaxometry capable o
 
 * **End-users using a GUI**: this mode is recommended for scientists who are not specialists in image processing or programming. Download FijiRelax through the official Fiji release, and follow the step-by-step installation instructions, as well as the hands-on tutorials built on the test dataset hosted at Zenodo [@fijirelaxDataset]. Then, use the graphical user interface to import and process your own Bruker/NIFTI/Custom data, explore the relaxation maps in space and time using the graphical relaxation curve explorer and export your results as 2D/3D/4D TIFF images. This mode is also recommended for studying new datasets or new biological questions. Among the interface features, the plugin provides a graphical explorer to visualize the relaxation curves, and the estimated PD-weighted T1 and T2 distributions over customizable areas of interest. In 5D hypermaps, the distributions at each timepoint can be displayed simultaneously, giving access to valuable information on water distribution in tissues and its evolution during the monitoring period. 
 
-* **Advanced users**: this mode can be used by scientists with programming skills. Load the sample BeanShell scripts provided in the test directory of the Github repository ([https://github.com/rocsg/fijirelax](https://github.com/rocsg/fijirelax)) by dragging them into the Fiji interface and run the scripts to reproduce the results shown in \autoref{fig:figure1}: import a dataset, convert it to a HyperMap (see \autoref{fig:figure1}-e), compute the parameter maps. Then, adapt these scripts to your needs to process your own data and batch-process multiple experiments.
+* **Advanced users**: this mode can be used by scientists with programming skills. Load the sample BeanShell scripts provided in the test directory of the GitHub repository ([https://github.com/rocsg/fijirelax](https://github.com/rocsg/fijirelax)) by dragging them into the Fiji interface and run the scripts to reproduce the results shown in \autoref{fig:figure1}: import a dataset, convert it to a HyperMap (see \autoref{fig:figure1}-e), compute the parameter maps. Then, adapt these scripts to your needs to process your own data and batch-process multiple experiments.
 
-* **Developers**: this mode is for programmers fluent with Java and Maven. Start by exploring the FijiRelax API: [API Overview](https://javadoc.io/doc/io.github.rocsg/fijirelax/latest/index.html). Build your own tools on top of the FijiRelax library, provided as a jar file hosted at Maven central repository ([Artifact](https://search.maven.org/artifact/io.github.rocsg/fijirelax)), by indicating FijiRelax as a dependency in your POM file and run the unit tests. FijiRelax is hosted on a Github public repository ([https://github.com/rocsg/fijirelax](https://github.com/rocsg/fijirelax)) and developers can offer to contribute to its development and extend it by requesting features, or proposing new features.
+* **Developers**: this mode is for programmers fluent with Java and Maven. Start by exploring the FijiRelax API: [API Overview](https://javadoc.io/doc/io.github.rocsg/fijirelax/latest/index.html). Build your own tools on top of the FijiRelax library, provided as a jar file hosted at Maven central repository ([Artifact](https://search.maven.org/artifact/io.github.rocsg/fijirelax)), by indicating FijiRelax as a dependency in your POM file and run the unit tests. FijiRelax is hosted on a GitHub public repository ([https://github.com/rocsg/fijirelax](https://github.com/rocsg/fijirelax)) and developers can offer to contribute to its development and extend it by requesting features, or proposing new features.
 
 # State of the field
[…]
