### 2023-01-29 #4981 @arfon
@Rocsg – could you please merge this pull request which fixes a couple of issues with your paper? https://github.com/Rocsg/FijiRelax/pull/20
