### 2023-01-30 #4981 @arfon
Looks like there's a minor issue with the GitHub Pages build (which serves the PDF from the JOSS site). We'll keep this review issue open until the Pages issue is resolved.
